Move core datasets to JSON files
