+++ .gitignore
@@ -8,3 +8,4 @@ node_modules
 !.env.example
 .prettierrc
 .netlify
+/static/data/*
