+++ src/data/interfaces/interfaces.ts
@@ -6,7 +6,7 @@ export type OutlineObject = {
 
 type LineDetails = {
 	path: string;
-	translateValues: [number, number];
+	translateValues: number[];
 };
 
 export interface SyllabusSection {
