--- src/data/outlines/alphabet.ts
@@ -1,183 +1,0 @@
-export const alphabet = [
-	{
-		name: 'a',
-		svgPaths: [
-			'M0 60.2682C4.24876 47.5219 14.8745 35.4645 20.7287 23.2155C21.0135 22.6197 29.8893 -0.875642 33.9433 0.0252572C35.7865 0.434854 67.3364 37.7487 75.2713 44.3329C77.7854 46.4191 89.3424 56.3974 87.9676 56.1224'
-		],
-		translateValues: '283.292847503344, 397.858683469488'
-	},
-	{
-		name: 'b',
-		svgPaths: [
-			'M74.798 0C70.4679 12.9902 55.6928 24.3647 48.3688 35.8866C26.6894 69.992 -1.2733 117.817 0.0449239 160C2.77571 247.385 80.8371 228.752 98.2474 171.401C111.841 126.623 14.6489 106.248 1.34047 159.482'
-		],
-		translateValues: '274.567761627583, 236.329284750337'
-	},
-	{
-		name: 'c',
-		svgPaths: [
-			'M58.6715 4.08549C58.6715 2.46593 56.0258 2.07375 54.7849 1.75351C27.3147 -5.33558 -3.75982 9.33192 0.371918 41.0086C3.05647 61.5901 18.6145 73.5526 38.461 60.8304C42.945 57.9561 47.5201 55.5844 52.3233 53.3163C53.7758 52.6304 58.0866 50.2711 57.8942 50.207'
-		],
-		translateValues: '297.949269202232, 390.55958156634'
-	},
-	{
-		name: 'd',
-		svgPaths: [
-			'M6.82189 0C-3.24926 2.18938 -2.06203 1.84418 9.67209 1.4251C31.6996 0.638404 53.731 0.29227 75.745 1.94332C86.8468 2.77596 98.3791 2.59109 109.429 2.59109C109.777 2.59109 108.741 2.72065 108.393 2.72065'
-		],
-		translateValues: '300.827230951816, 457.219973009447'
-	},
-	{
-		name: 'e',
-		svgPaths: [
-			'M0 1.58539e-05C0 -0.0161497 3.25077 12.3348 3.36842 15.8057C3.72459 26.3126 5.50441 37.0311 3.88664 47.5466C3.68471 48.8591 0.915179 58.7164 1.81377 59.2065C4.98583 60.9367 55.8191 59.4887 65.8138 60.1134C68.4376 60.2774 70.1017 59.7057 68.2753 62.4454'
-		],
-		translateValues: '292.232118758404, 391.794855940966'
-	},
-	{
-		name: 'f',
-		svgPaths: [
-			'M2.61941 166.915C2.1895 166.915 2.06478 168.136 3.39674 166.656C12.6902 156.33 23.5085 144.022 30.7328 132.065C39.716 117.196 107.662 14.9276 71.672 1.08516C37.349 -12.116 11.7122 99.1004 6.3765 119.109C2.75061 132.707 2.48986 146.603 2.48986 160.437C2.48986 163.756 -1.71682 172.219 0.805647 167.174'
-		],
-		translateValues: '298.681532305607, 288.750193081841'
-	},
-	{
-		name: 'g',
-		svgPaths: [
-			'M3.17134 22.6047C-2.12471 22.1634 28.0044 -1.36271 40.3535 0.0621974C65.0871 2.91607 71.4982 18.5035 72.3535 40.7424C73.7586 77.2741 42.8407 115.096 25.0661 144.386C18.0119 156.01 11.9221 168.059 5.24421 179.884C4.39668 181.385 -0.611003 194.102 0.0620246 186.362'
-		],
-		translateValues: '308.493980748327, 385.514050949693'
-	},
-	{
-		name: 'h',
-		svgPaths: ['M2.46154 0C8.0645 16.8089 0 158.953 0 196.016'],
-		translateValues: '325.527665317109, 262.62887989203'
-	},
-	{
-		name: 'i',
-		svgPaths: [
-			'M28.4331 0C30.553 1.4133 25.1201 6.78327 24.5465 7.64372C17.4059 18.3546 10.2725 28.9518 2.91084 39.5142C-10.7314 59.0879 27.8155 24.2879 31.2833 22.1538C40.5342 16.461 49.6972 11.2542 57.4533 3.49798'
-		],
-		translateValues: '347.232213316981, 411.61673414305'
-	},
-	{
-		name: 'j',
-		svgPaths: [
-			'M107.919 0.642606C107.919 -2.51698 106.65 6.83834 106.235 9.97054C104.807 20.7439 106.217 31.8003 105.717 42.6183C104.403 70.9912 103.825 100.165 100.405 128.383C96.7319 158.685 89.3711 208.25 65.9433 231.12C52.6349 244.112 36.5294 252.595 19.5628 259.234C16.7182 260.347 0.504738 258.122 0 260.141'
-		],
-		translateValues: '289.122807017514, 334.53688095716'
-	},
-	{
-		name: 'k',
-		svgPaths: [
-			'M43.179 0.684699C44.7946 -2.00801 40.6329 4.03878 39.4219 5.08956C32.659 10.9574 1.39402 31.3493 0.0372553 40.8466C-0.659749 45.7257 8.60028 45.7367 12.9927 47.9721C23.4798 53.3093 33.0897 59.9886 43.3085 65.7211C45.1999 66.7822 54.3741 68.1827 49.0089 68.1827'
-		],
-		translateValues: '337.409438398721, 388.778188570915'
-	},
-	{
-		name: 'l',
-		svgPaths: [
-			'M69.2485 0.327705C72.1617 -3.70596 42.8325 30.6871 37.2485 38.1577C16.2587 66.2385 1.71265 105.214 0.066286 140.117C-1.67057 176.939 30.9692 224.8 69.2485 193.882C70.5925 192.797 83.7117 184.849 81.5562 180.538'
-		],
-		translateValues: '307.712391474792, 248.050163145801'
-	},
-	{
-		name: 'm',
-		svgPaths: [
-			'M0.382651 30.7188C-3.7565 30.7188 26.9068 13.4188 33.4191 11.156C58.7815 2.34367 85.9006 -3.37205 112.447 2.21675C116.657 3.10299 155.888 25.2944 151.443 28.2572'
-		],
-		translateValues: '251.039750699639, 428.57406895552'
-	},
-	{
-		name: 'n',
-		svgPaths: [
-			'M0.159054 13.8725C-1.41044 13.3493 9.14418 4.28474 9.48699 4.02631C21.5392 -5.05921 28.8456 2.5343 34.3615 13.8725C40.4194 26.3249 36.8725 47.172 36.3048 60.5121C35.801 72.3504 36.8564 83.0587 36.823 82.9251'
-		],
-		translateValues: '341.044724257517, 369.371802485656'
-	},
-	{
-		name: 'o',
-		svgPaths: [
-			'M0.429231 18.0749C0.567341 17.8677 0.128179 25.5667 0.0405675 27.1437C-0.317389 33.5869 1.59082 45.1513 9.88672 46.1883C25.0307 48.0813 45.7798 33.3513 48.2349 18.0749C52.8101 -10.3931 -2.02838 -2.6224 1.20656 23.2571'
-		],
-		translateValues: '321.600458113521, 408.699718585701'
-	},
-	{
-		name: 'p',
-		svgPaths: [
-			'M2.91741 0.0561427C1.2198 -0.509728 1.41686 3.34696 1.10365 5.10877C-1.72955 21.0455 1.73927 40.1208 2.14009 56.1533C3.17248 97.449 9.61097 136.235 7.97005 177.805C7.12835 199.128 7.38811 219.553 8.0996 240.898C8.11979 241.504 7.97005 243.318 7.97005 242.712'
-		],
-		translateValues: '323.258024592235, 340.69419464797'
-	},
-	{
-		name: 'q',
-		svgPaths: [
-			'M28.0655 2.07287C30.3589 2.07287 23.5694 14.1622 23.0129 15.417C17.305 28.2856 14.4203 42.4492 10.5756 55.9676C2.86648 83.0739 -14.2874 162.058 23.9197 175.158C58.2442 186.926 86.5571 44.2607 87.272 18.5263C87.3057 17.3106 88.5096 0.536126 86.3651 0'
-		],
-		translateValues: '279.583640437522, 383.892037786774'
-	},
-	{
-		name: 'r',
-		svgPaths: [
-			'M0.0175922 182.543C-0.689805 183.604 20.0661 164.256 38.884 140.826C64.3145 109.162 88.9611 77.9544 115.839 47.5466C128.695 33.0027 142.295 19.2599 155.095 4.66397C156.393 3.18315 160.692 0 158.722 0'
-		],
-		translateValues: '273.169992144249, 272.345479082321'
-	},
-	{
-		name: 's',
-		svgPaths: [
-			'M3.55151 21.5289C1.43854 21.5289 4.2988 26.4856 4.7175 27.0997C12.0785 37.8959 41.0134 23.5474 36.0697 11.6827C23.6514 -18.1212 -12.3428 18.0088 4.32884 22.1767'
-		],
-		translateValues: '324.826357923219, 422.476514101815'
-	},
-	{
-		name: 't',
-		svgPaths: [
-			'M3.11026 6.11941C-6.51375 6.11941 9.13197 4.82471 10.6244 4.6943C35.0405 2.56086 59.2508 1.65983 83.6933 0.0303372C87.2889 -0.209374 107.886 1.02922 103.774 1.71455'
-		],
-		translateValues: '286.789872819412, 307.16534435215'
-	},
-	{
-		name: 'u',
-		svgPaths: [
-			'M3.31531 1.07448C3.31531 -4.17671 0.410944 11.251 0.0764433 16.4915C-0.519162 25.8226 2.20131 40.9541 14.0683 40.3295C27.428 39.6264 31.0542 23.2412 30.0036 12.7344C29.8592 11.2905 28.0602 -1.07671 28.0602 2.49958'
-		],
-		translateValues: '327.653651157309, 414.558451478843'
-	},
-	{
-		name: 'v',
-		svgPaths: [
-			'M2.20912 5.47028C-8.15356 2.01605 20.952 39.4146 27.3427 49.0006C27.7215 49.5688 28.3431 51.3473 28.5087 50.6849C32.5792 34.4031 43.0715 15.95 52.4763 1.84275C53.6505 0.0815513 54.3849 -1.05021 53.3832 1.45409'
-		],
-		translateValues: '316.840808888024, 407.441998506733'
-	},
-	{
-		name: 'w',
-		svgPaths: [
-			'M0.232654 5.81272C-0.890153 4.68991 2.34915 8.20437 3.60108 9.18114C22.2482 23.73 46.0876 41.5951 70.9695 40.7925C91.3189 40.136 110.284 28.9429 127.844 19.8046C135.381 15.8825 142.577 12.2824 148.314 5.94228C149.089 5.08503 154.476 -1.31058 151.682 0.241871'
-		],
-		translateValues: '257.667480509105, 412.411300370559'
-	},
-	{
-		name: 'x',
-		svgPaths: [
-			'M0 0C3.98232 1.32744 30.7045 32.5453 30.7045 39.1255',
-			'M0.0252486 47.7233C-0.535982 48.0974 8.4237 36.6571 10.5192 34.6382C19.2998 26.1788 28.717 18.6191 36.8188 9.50464C37.9764 8.20233 48.0929 -2.55556 47.3127 0.565374'
-		],
-		translateValues: '330.968960863668, 413.300944669366'
-	},
-	{
-		name: 'y',
-		svgPaths: [
-			'M0 4.79352C5.40523 12.9014 -8.68015 55.1795 13.3441 56.0972C30.5332 56.8134 33.9433 15.6662 33.9433 5.18219C33.9433 4.92405 33.9609 0 33.8138 0C32.7125 0 36.852 35.6286 37.5709 40.4211C38.2344 44.8447 42.4834 57.6201 41.0688 53.3765'
-		],
-		translateValues: '301.171390013465, 399.827260458809'
-	},
-	{
-		name: 'z',
-		svgPaths: [
-			'M10.8583 1.55817C6.98705 2.8486 18.7223 6.93598 18.6316 11.0157C17.8684 45.3578 -26.4321 1.95598 24.332 0.00351658C27.3101 -0.111027 29.9287 2.59448 31.328 4.92659C39.8222 19.0837 26.2393 38.6317 17.9838 49.3638C14.0518 54.4755 3.12713 69.8157 4.89881 63.6149'
-		],
-		translateValues: '320.887959640116, 407.337913915582'
-	}
-];
