--- src/data/outlines/outlines.ts
@@ -1,2495 +1,0 @@
-import type { OutlineObject } from '../interfaces/interfaces';
-
-export const outlines: OutlineObject[] = [
-	{
-		specialOutlineMeanings: ['a', 'able', 'able to', 'ability', 'after'],
-		lines: [
-			{
-				path: 'M0 60.2682C4.24876 47.5219 14.8745 35.4645 20.7287 23.2155C21.0135 22.6197 29.8893 -0.875642 33.9433 0.0252572C35.7865 0.434854 67.3364 37.7487 75.2713 44.3329C77.7854 46.4191 89.3424 56.3974 87.9676 56.1224',
-				translateValues: [310, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['a'],
-		lines: [
-			{
-				path: 'M0 0.109179C2.97633 -2.12307 32.4813 30.4077 50.9305 48.857',
-				translateValues: [330, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['a lot of'],
-		lines: [
-			{
-				path: 'M67.9773 0C68.3252 0.731233 68.137 1.09518 67.6944 1.78439C45.7294 35.987 -8.74936 109.231 1.20177 153.061C14.1948 210.288 98.2527 166.115 98.2527 157.819',
-				translateValues: [270, 460]
-			}
-		]
-	},
-	{
-		letterGroupings: ['abs'],
-		specialOutlineMeanings: ['absolute'],
-		lines: [
-			{
-				path: 'M51.4211 0C51.4211 10.3889 67.9808 25.4685 76.3205 33.1103C76.4578 33.2361 84.1606 39.7108 83.96 39.8513C61.285 55.74 41.2301 71.0727 25.107 94.3743C1.93724 127.86 -21.4967 227.747 42.9327 232.763C71.458 234.984 125.255 183.179 100.371 153.854C63.0701 109.896 -14.8289 144.875 2.4712 169.12C21.6204 195.957 47.4087 155.486 25.3899 146.915C23.1701 146.05 10.1108 143.473 10.1108 147.509',
-				translateValues: [300, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['advantage'],
-		lines: [
-			{
-				path: 'M0 0C3.21745 6.76354 13.7808 12.4877 19.2404 18.4387C22.0222 21.4708 34.3236 47.5837 36.7832 47.5837C59.3957 47.5837 100.687 37.5336 122.233 46.5923C135.391 52.1241 146.713 83.3225 160.714 88.228C163.014 89.034 163.31 84.1011 164.675 82.0818C176.753 64.2149 221.6 20.0849 230.885 65.6258C233.175 76.8573 233.715 88.1635 233.715 99.7274',
-				translateValues: [220, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['advert'],
-		lines: [
-			{
-				path: 'M0 0.694519C1.43141 -2.3145 4.938 5.26076 6.79074 8.03033C11.5244 15.1065 27.8542 51.4172 36.2173 55.8123C60.2055 68.4189 123.533 46.5324 144.586 57.5966C151.764 61.3691 165.165 101.017 176.559 101.017C178.818 101.017 203.704 74.6964 205.986 72.07C208.736 68.9041 220.632 59.8242 219.001 60.9672',
-				translateValues: [230, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['after all'],
-		lines: [
-			{
-				path: 'M0.106431 175.687C-1.97445 180.547 27.0416 131.49 36.0407 127.707C38.2436 126.78 73.9795 172.911 76.5022 172.911C76.6458 172.911 74.0502 163.059 74.2386 159.429C74.8945 146.79 74.6917 134.235 79.6146 122.353C93.8671 87.9541 115.696 66.3458 143.561 42.2542C162.937 25.502 187.09 10.944 211.468 2.40292C215.555 0.970888 237.868 -1.62607 227.03 1.4116',
-				translateValues: [250, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['air ', 'arch'],
-		lines: [
-			{
-				path: 'M0.0327279 199.851C-0.750222 199.668 12.7541 182.689 14.8875 179.777C25.8911 164.755 36.8575 150.218 46.719 134.424C54.2016 122.44 149.39 0 156.22 0C159.324 0 160.501 4.5137 162.374 6.98885C191.204 45.0835 209.602 95.1574 234.101 136.357C246.098 156.532 261.621 170.673 276.967 187.955C278.603 189.797 288.593 200.573 286.729 199.703',
-				translateValues: [220, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['already'],
-		lines: [
-			{
-				path: 'M121.596 0C124.495 0 153.809 25.678 150.173 30.1363C79.8311 116.394 -4.78591 198.517 0.211305 320.198C1.03056 340.147 11.1326 370.016 27.9401 382.255C104.609 438.085 121.227 357.45 135.743 338.835C155.027 338.648 197.285 331.764 214.968 340.025C216.215 340.607 205.51 350.34 188.937 364.015',
-				translateValues: [260, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['also'],
-		lines: [
-			{
-				path: 'M3.24044 12.0366C3.24044 29.2928 46.5401 13.5507 34.3646 3.31289C17.3063 -11.0308 -17.2621 25.1269 10.3141 37.018C25.1505 43.4156 43.4961 38.9981 52.4733 25.5186C54.7689 22.0716 59.009 5.56362 57.5663 7.07993',
-				translateValues: [330, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['alternative'],
-		lines: [
-			{
-				path: 'M64.6838 0C68.9854 4.52123 90.432 25.0176 90.432 31.9207C90.432 39.1263 54.2684 53.4816 48.2729 59.083C13.8568 91.2375 -23.8344 155.878 19.4123 196.283C42.5908 217.938 79.6265 226.118 102.033 196.679C104.458 193.493 110.156 180.054 113.068 179.034C137.304 177.024 151.33 176.107 167.303 178.338C183.617 176.274 191.882 207.855 198.983 221.538C200.833 225.103 206.135 213.203 208.583 210.018C220.554 194.439 243.266 170.173 238.131 174.672',
-				translateValues: [250, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['always'],
-		lines: [
-			{
-				path: 'M0.286138 9.47079C-1.43337 9.47079 4.07781 20.238 31.087 34.3297C54.4334 46.5105 85.9382 53.4992 112.091 48.3963C121.593 46.5422 176.582 22.962 165.689 7.53058C144.465 -22.5367 122.239 48.4621 151.623 28.8729',
-				translateValues: [240, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['amount'],
-		lines: [
-			{
-				path: 'M0 0.589886C0 6.66283 8.78034 14.4197 11.8838 19.2268C14.5492 23.3555 18.0492 38.7612 22.0699 40.6395C26.0863 42.5157 62.2611 14.7372 65.6438 13.0806C104.536 -5.96678 179.516 -9.25339 202.024 38.062',
-				translateValues: [230, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['and the'],
-		lines: [
-			{
-				path: 'M0 6.68359C2.99254 9.82896 21.845 -13.3654 28.4362 12.0368C32.8526 29.0571 29.6799 47.9247 27.1629 64.9736C26.4532 69.7814 19.1522 77.0431 23.5554 79.1C35.4257 84.6451 84.0452 72.9307 100.588 75.8286C107.571 77.0519 99.708 90.1242 101.012 97.0925C108.671 138.015 116.504 205.526 116.504 246.535',
-				translateValues: [300, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['anything'],
-		lines: [
-			{
-				path: 'M0.231166 0.00171567C-2.65984 -0.287385 22.512 36.0579 22.968 34.69C26.5547 23.9298 47.1213 6.22112 56.9761 20.0179C64.6685 30.7872 62.3451 53.7206 61.0571 66.1718C56.8436 106.902 73.1194 64.6068 89.0409 52.6657',
-				translateValues: [290, 370]
-			},
-			{
-				path: 'M0 46.8566C0 41.7265 17.0245 18.3789 19.4332 14.1117C22.1285 9.3367 29.7328 -4.19739 29.7328 1.28577',
-				translateValues: [390, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['appropriate'],
-		lines: [
-			{
-				path: 'M0 0C3.5956 3.77923 44.3113 23.8161 44.7057 29.3433C46.297 51.644 37.6692 74.975 39.0467 98.1413C43.8277 178.543 40.7444 260.443 40.7444 340.818',
-				translateValues: [280, 260]
-			},
-			{
-				path: 'M1.92331 0C0.271881 15.568 1.07447 56.7181 1.07447 71.9703C1.07447 93.7435 4.20194 115.47 3.22689 138.017C1.83927 170.104 -3.18428 206.848 3.0551 238.91C2.4404 270.976 6.90291 278.371 4.18689 300.257C4.00247 301.743 3.46404 309.626 5.03573 310.483C25.1131 314.593 72.3469 308.065 84.8269 309.492',
-				translateValues: [370, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['approximately'],
-		lines: [
-			{
-				path: 'M0 0C0 7.23385 28.621 20.0003 33.6194 27.498C35.6217 30.5014 29.0329 192.51 32.8421 216C36.7411 240.044 33.1656 270.289 31.4818 294.704C30.9428 302.52 29.9271 310.045 29.9271 317.927C29.9271 323.507 54.3577 287.52 61.2146 274.785C92.4623 216.754 130.196 177.067 168.486 123.984C182.378 104.725 186.608 82.3473 203.466 65.4899',
-				translateValues: [240, 260]
-			},
-			{
-				path: 'M0.753221 0C-5.08722 0 24.4812 14.779 39.6196 30.3158C42.8559 33.6372 54.6878 40.2118 50.1135 39.4494',
-				translateValues: [390, 340]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['area'],
-		lines: [
-			{
-				path: 'M0 86.8567C12.6448 83.9033 44.6792 7.92887 58.8741 0.192641C61.2127 -1.08186 62.2613 4.31469 64.1465 6.19601C69.5225 11.5612 134.43 83.2616 137.3 83.932',
-				translateValues: [250, 370]
-			},
-			{
-				path: 'M1.69457e-06 89.5888C-0.00758315 89.587 25.4492 60.4053 30.5355 55.4158C45.3701 40.863 58.5914 24.7175 72.714 9.54382C74.2383 7.90605 79.1087 0 81.2815 0C84.1801 0 94.9115 12.9995 96.2197 14.3157C119.494 37.7332 145.947 61.2722 168.055 84.509',
-				translateValues: [320, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['arrange'],
-		lines: [
-			{
-				path: 'M0.215366 0C-1.53009 0 7.8764 14.2383 8.916 16.0595C15.9541 28.3887 23.6629 40.8142 29.7126 53.5316',
-				translateValues: [320, 400]
-			},
-			{
-				path: 'M0 0C13.6605 14.3581 24.2135 42.808 33.317 60.6692C33.967 61.9444 34.6518 63.6627 35.0147 62.8996',
-				translateValues: [350, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['as a result'],
-		lines: [
-			{
-				path: 'M0 0C0 10.9593 15.6415 30.7949 21.7869 40.8426C23.0181 42.8555 27.9188 47.7354 24.0505 45.7993',
-				translateValues: [270, 410]
-			},
-			{
-				path: 'M0 0.148026C0.582341 -1.07614 4.15763 5.64691 5.09305 7.28558C10.9895 17.6148 15.5331 29.6082 19.5234 40.7924C20.9283 44.7303 25.8466 56.2657 24.0505 52.4901',
-				translateValues: [310, 400]
-			},
-			{
-				path: 'M0.463143 205.21C1.27021 206.059 -2.48242 208.903 3.29262 199.461C12.9946 183.598 28.7367 171.137 39.7928 156.239C63.2577 124.619 85.8998 87.3338 104.305 52.9426C113.311 36.1146 125.517 22.9575 135.995 7.14335C136.675 6.11734 142.787 -2.97121 141.371 0.997123',
-				translateValues: [370, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['association'],
-		lines: [
-			{
-				path: 'M0 0C62.6166 65.8145 19.3334 51.3869 18.6745 47.3854C16.2507 32.666 46.1209 13.7803 54.6088 30.1363C63.8056 47.8581 56.8724 74.1935 56.8724 93.1847',
-				translateValues: [290, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['at'],
-		lines: [
-			{
-				path: 'M0 0.0729122C10.697 -1.45523 31.9697 21.457 39.0467 29.4186C40.3696 30.9068 51.0559 42.7369 51.6581 39.726',
-				translateValues: [330, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['at the'],
-		lines: [
-			{
-				path: 'M12.4138 0C22.7799 6.97188 25.2458 10.3827 28.3505 12.1957C36.7582 17.1051 57.1844 28.2896 65.2567 39.9665C55.9326 57.6905 24.9717 140.756 20.5919 166.331C19.8386 170.73 18.8339 174.557 17.6561 178.821C12.9108 196 8.58102 213.119 4.65509 230.542C3.67732 234.881 -0.458567 249.841 0.0418216 249.35',
-				translateValues: [350, 210]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['at the same time'],
-		lines: [
-			{
-				path: 'M0 0C7.70567 3.59965 37.176 12.4464 42.1592 19.43C43.455 21.246 33.1048 134.975 33.1048 150.682C33.1048 167.787 40.8293 220.215 17.2598 226.221C12.8637 227.341 8.83472 223.217 7.35663 219.281C-4.85122 186.775 78.3935 170.749 99.0316 176.258C116.999 181.054 132.624 191.458 146.85 203.42C149.897 205.983 163.496 222.173 164.392 221.859C203.06 208.311 213.31 177.447 258.189 179.194C275.688 179.875 326.416 193.211 339.254 230.384',
-				translateValues: [190, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['atomic'],
-		lines: [
-			{
-				path: 'M0 0C0 4.71113 11.7869 13.4132 14.9962 16.8525C21.4611 23.7808 26.7483 31.4128 33.1048 38.2652C33.3776 38.5593 52.6699 25.3266 79.2253 10.3098C101.55 -2.31487 138.154 -3.25971 161.28 8.52541C166.785 11.331 172.129 14.594 177.125 18.2404C212.952 44.3907 160.044 23.8132 146.284 52.7386C135.831 74.7116 173.92 79.1063 185.613 66.8154',
-				translateValues: [240, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['attention'],
-		lines: [
-			{
-				path: 'M0 0C6.12029 9.18043 11.796 19.107 20.1297 27.1629C34.5809 41.1324 24.5878 39.3178 36.6215 27.2842C43.7028 20.2029 59.1904 5.47033 69.3625 15.2792C83.9375 29.3336 72.4872 69.1446 70.5752 87.3095C70.4291 88.6965 69.7111 97.6016 69.12 92.2813',
-				translateValues: [320, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['auto'],
-		lines: [
-			{
-				path: 'M0 52.342C4.12243 45.1204 26.8316 0 37.0661 0C39.2162 0 45.3816 13.8635 46.6863 15.8612C54.1339 27.2649 62.2237 34.1034 72.7175 42.2305C75.1478 44.1127 82.4597 48.1215 78.9423 49.7646',
-				translateValues: [310, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['b', 'be', 'been'],
-		lines: [
-			{
-				path: 'M74.798 0C70.4679 12.9902 55.6928 24.3647 48.3688 35.8866C26.6894 69.992 -1.2733 117.817 0.0449239 160C2.77571 247.385 80.8371 228.752 98.2474 171.401C111.841 126.623 14.6489 106.248 1.34047 159.482',
-				translateValues: [320, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['because'],
-		lines: [
-			{
-				path: 'M26.0466 0C26.7284 1.43309 13.7279 9.24376 11.5777 11.7549C2.98277 21.7926 -1.09811 37.4261 0.25426 50.6928C2.57951 73.5036 49.6394 58.2646 42.4028 43.0522C31.8364 20.84 0.834366 63.5856 30.8696 56.5703',
-				translateValues: [320, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['begin', 'began', 'begun'],
-		lines: [
-			{
-				path: 'M0.374631 25.8929C-3.19851 40.9154 19.4385 -17.6461 40.5532 5.47161C58.0997 24.6827 43.9485 75.4864 43.9485 100.441',
-				translateValues: [320, 460]
-			}
-		]
-	},
-	{
-		letterGroupings: ['bs'],
-		specialOutlineMeanings: ['business'],
-		lines: [
-			{
-				path: 'M73.9903 0.141252C77.8059 -1.76656 66.4191 16.2609 65.7444 17.4819C56.8512 33.5744 46.9503 48.4566 37.6113 64.4107C11.8153 108.479 -27.5222 181.648 29.3654 217.202C71.6448 243.627 140.738 213.965 122.738 155.964C109.518 113.368 13.4355 122.995 1.47491 163.361C-11.3841 206.761 65.01 161.5 31.0631 151.599C22.1821 149.009 11.1262 151.394 3.90018 156.813',
-				translateValues: [300, 230]
-			}
-		]
-	},
-	{
-		letterGroupings: ['bp', 'bh'],
-		specialOutlineMeanings: ['by the'],
-		lines: [
-			{
-				path: 'M63.8439 0C66.0349 6.14099 -26.854 136.698 40.0763 190.632C112.923 249.333 159.732 111.529 71.6957 123.866C-0.792002 134.025 9.50868 238.585 2.72727 290.855C1.48947 300.396 -2.96674 364.073 3.15169 365.502',
-				translateValues: [300, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['c', 'once', 'offence'],
-		lines: [
-			{
-				path: 'M58.6715 4.08549C58.6715 2.46593 56.0258 2.07375 54.7849 1.75351C27.3147 -5.33558 -3.75982 9.33192 0.371918 41.0086C3.05647 61.5901 18.6145 73.5526 38.461 60.8304C42.945 57.9561 47.5201 55.5844 52.3233 53.3163C53.7758 52.6304 58.0866 50.2711 57.8942 50.207',
-				translateValues: [300, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['chairman'],
-		lines: [
-			{
-				path: 'M34.9437 1.66079C39.1963 0.667509 26.3023 -0.210073 21.9427 0.0444975C11.9313 0.629092 -2.40646 12.7244 0.344173 24.2889C2.29203 32.4783 7.85507 37.4484 15.6518 39.1294C20.9892 40.2801 32.008 32.7878 32.008 38.2478C32.008 99.9893 13.1355 163.44 13.1355 225.444',
-				translateValues: [340, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['child'],
-		lines: [
-			{
-				path: 'M37.5691 0.676771C39.0867 -0.741088 23.7292 0.477719 22.4711 0.676771C12.6081 2.23736 0.533499 10.5415 0.0338616 21.3947C-1.08184 45.6302 25.6911 30.6723 33.3752 37.8515C33.4253 37.8983 25.8015 112.6 29.8104 133.066C34.9396 159.251 25.0873 184.671 18.6966 209.619C18.4645 210.526 17.0866 218.189 17.8579 218.729C27.9337 225.79 75.1686 223.04 90.2024 220.933C91.335 220.775 96.4276 219.032 96.9126 220.052',
-				translateValues: [290, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['children'],
-		lines: [
-			{
-				path: 'M37.9801 0.121522C39.4812 -0.579709 24.7112 1.93224 21.8521 3.0955C11.2673 7.40212 -1.00308 16.7397 0.0651633 29.4648C2.35917 56.7913 40.4302 20.227 39.3948 59.4028C38.681 86.4139 34.0666 113.488 32.0382 140.493C29.8359 169.814 36.1552 195.148 29.7746 224.954C29.4567 226.44 27.483 230.085 28.6428 231.101C54.7388 253.958 258.759 206.161 278.202 248.548C286.05 265.655 278.202 283.282 278.202 300.89',
-				translateValues: [190, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['church'],
-		lines: [
-			{
-				path: 'M1.8203 174.563C-1.05608 184.631 0.232331 138.41 0.65431 131.94C2.70758 100.456 7.74742 70.033 12.4438 38.9196C14.0549 28.246 15.2911 17.4628 14.9053 6.66051C14.8629 5.47152 12.9852 -3.30186 12.0551 1.34877',
-				translateValues: [350, 280]
-			},
-			{
-				path: 'M0.102946 7.35409C-2.07159 8.07893 30.8279 3.10051 39.7467 2.04235C41.1195 1.87947 51.7817 -1.15916 50.3702 0.48769',
-				translateValues: [340, 330]
-			}
-		]
-	},
-	{
-		letterGroupings: ['crcs'],
-		specialOutlineMeanings: ['circumstances'],
-		lines: [
-			{
-				path: 'M34.2875 122.082C-11.0259 100.914 -12.3275 193.193 35.773 160.892C97.8686 119.194 146.673 54.278 202.146 4.60967C202.549 4.24847 205.66 0 206.602 0C212.667 0 197.676 8.21353 193.233 12.342C185.917 19.1403 170.639 36.6105 177.105 47.7324C187.683 65.9248 227.743 47.0054 222.306 27.9554C220.203 20.5865 209.708 22.438 204.905 24.3866C176.675 35.8388 214.454 45.8275 214.454 40.7435',
-				translateValues: [270, 290]
-			}
-		]
-	},
-	{
-		letterGroupings: ['crc'],
-		specialOutlineMeanings: ['circumstance'],
-		lines: [
-			{
-				path: 'M25.8197 137.407C-9.23652 137.407 -9.22069 186.955 29.7064 176.079C51.0628 170.111 64.5879 152.715 81.2043 139.35C112.617 114.083 239.091 8.20575 267.763 0.0138116C268.621 -0.231187 264.383 2.74677 250.273 11.8681C239.753 18.6687 219.921 31.7687 222.095 47.528C224.287 63.4195 260.541 49.3966 267.374 44.5158',
-				translateValues: [220, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['city'],
-		lines: [
-			{
-				path: 'M52.6742 0.76542C16.0828 -5.6446 -25.5767 30.0255 19.9619 39.9973C30.0274 42.2013 39.3936 37.1483 48.69 38.234C51.06 38.5108 45.439 41.7286 43.867 43.5237C40.6495 47.1978 31.7047 56.5947 31.7047 62.7723',
-				translateValues: [320, 240]
-			}
-		]
-	},
-	{
-		letterGroupings: ['cb'],
-		specialOutlineMeanings: ['club'],
-		lines: [
-			{
-				path: 'M73.8438 2.59117C57.494 -5.04651 27.4548 4.67635 28.6429 25.4908C29.457 39.7516 46.5121 42.8355 57.7158 40.5094C60.57 39.9168 65.992 34.7691 65.992 37.6841C65.992 40.3655 61.4836 40.589 59.2013 41.9964C25.2302 62.9439 -5.53066 106.034 0.84337 148.465C7.93733 195.688 81.727 223.733 102.917 168.39C123.102 115.67 28.6348 102.385 8.69516 144.301C7.82543 146.13 4.3342 165.394 8.69516 164.376',
-				translateValues: [310, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['committee'],
-		lines: [
-			{
-				path: 'M35.1506 0C37.0173 1.96205 31.837 4.29686 30.9064 5.1549C24.372 11.1795 1.91104 37.9346 0.0651082 46.9889C-0.953672 51.986 10.2451 46.2851 15.3443 46.1958C30.4729 45.9308 130.95 42.9334 137.295 45.601C139.301 46.4446 136.779 49.1968 136.446 50.7559C133.951 62.4087 131.07 72.0887 131.07 84.0645C131.07 87.0483 127.881 104.199 130.585 102.936',
-				translateValues: [280, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['communication'],
-		lines: [
-			{
-				path: 'M271.868 5.39482C261.092 3.81253 240.997 2.70606 238.268 2.51482C214.899 0.877369 204.102 1.69207 180.708 0.287012C138.079 -2.27331 -5.90295 11.4594 0.187537 75.4748C5.6529 132.92 157 126.994 195.987 120.878C202.401 119.871 268.422 113.664 268.422 111.91C268.422 110.611 248.183 115.241 245.948 115.795C231.844 119.32 226.05 114.751 220.988 143.635C230.303 161.477 257.886 157.424 267.29 153.546',
-				translateValues: [190, 340]
-			},
-			{
-				path: 'M0 17.9296C0.657507 9.11874 25.5373 -8.53702 35.0855 4.84409C44.4424 17.9571 40.4131 33.517 39.6126 48.6607C39.3713 53.2258 40.9987 59.5221 39.8956 66.9011',
-				translateValues: [470, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['community'],
-		lines: [
-			{
-				path: 'M231.229 16.4307C230.589 17.7767 217.737 10.9304 216.587 10.4827C200.971 4.40363 181.843 1.85227 165.232 0.668593C114.244 -2.96476 14.9287 6.51282 1.19315 71.7467C-15.9291 153.065 156.262 137.564 195.154 131.226C206.239 129.42 234.713 115.591 245.023 117.397C245.761 117.527 241.292 121.028 228.258 134.795C219.185 144.379 210.65 155.132 203.43 166.171C202.797 167.137 200.459 169.11 200.459 167.955',
-				translateValues: [240, 320]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['conference'],
-		lines: [
-			{
-				path: 'M64.1865 13.592C69.2904 16.4531 63.3151 5.92751 60.3667 4.07527C35.117 -11.7873 32.4102 22.662 33.416 38.8708C36.4078 87.0892 59.5854 188.931 21.1078 226.678C13.3863 234.252 4.43081 213.236 3.06987 208.536C-10.1521 162.874 22.639 123.219 38.7212 83.7779',
-				translateValues: [320, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['congratulate'],
-		lines: [
-			{
-				path: 'M51.8219 25.2169C54.9148 6.97688 31.2027 -15.2725 10.7642 14.6726C-9.28815 44.0519 15.4311 91.1074 0.0121864 123.52C-0.246516 124.064 3.69818 115.086 3.82187 114.497C6.83266 100.151 10.124 91.7423 18.4038 79.7036C32.1339 59.7403 73.2584 54.7202 89.7065 73.1609C113.532 99.8726 40.9164 244.997 19.8185 268.65',
-				translateValues: [320, 330]
-			},
-			{
-				path: 'M0.148119 0.564879C-0.943104 -1.72903 4.27427 3.64668 6.65591 4.53018C28.7896 12.741 55.6137 4.93025 77.3927 14.6417C78.7599 15.2513 78.1117 15.815 77.3927 16.8226',
-				translateValues: [350, 520]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['consequent'],
-		lines: [
-			{
-				path: 'M81.3957 11.2806C82.4466 10.1761 80.3027 8.18011 77.7174 5.7292C44.1681 -26.0763 34.4705 84.5023 40.9342 93.5607C55.7522 114.327 78.1353 66.4658 44.3296 81.8631C39.7388 83.954 35.6988 86.0181 32.1629 89.7937C21.8573 100.797 13.8229 113.07 7.82939 126.869C-3.21627 152.3 -5.83914 258.426 21.9768 273.586C88.2623 309.711 135.297 144.911 116.764 92.9659C116.678 92.7256 114.802 90.7586 115.349 89.9919C120.291 83.0665 177.629 82.7211 177.881 83.2509',
-				translateValues: [270, 360]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['convenient'],
-		lines: [
-			{
-				path: 'M41.9015 8.00801C44.2228 7.46581 32.1005 2.27088 30.5836 1.66352C21.1146 -2.12797 1.59065 0.152468 0.0253015 12.7664C-1.26844 23.1916 47.461 79.3835 52.3706 79.3835C53.6737 79.3835 53.5354 77.0292 54.3512 76.013C69.2577 57.4438 89.9132 46.5717 105.565 29.0241C109.409 24.7139 127.737 7.83181 119.995 12.1716',
-				translateValues: [280, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['country'],
-		lines: [
-			{
-				path: 'M52.95 16.2819C58.0885 16.2819 50.4909 11.489 48.9887 10.1357C39.852 1.9042 19.8159 -4.84138 8.81022 4.58423C-11.3377 21.8394 9.37611 66.6208 9.37611 87.4591',
-				translateValues: [340, 260]
-			}
-		]
-	},
-	{
-		letterGroupings: ['cn'],
-		specialOutlineMeanings: ['county', 'council'],
-		lines: [
-			{
-				path: 'M36.384 9.84266C39.3143 9.84266 32.4999 2.34742 31.2909 1.96055C22.1229 -0.973218 11.2677 -1.55081 5.09809 6.93234C-8.2271 25.2545 9.69131 54.8719 1.94525 76.1736',
-				translateValues: [340, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['d', 'do', 'day'],
-		lines: [
-			{
-				path: 'M6.82189 0C-3.24926 2.18938 -2.06203 1.84418 9.67209 1.4251C31.6996 0.638404 53.731 0.29227 75.745 1.94332C86.8468 2.77596 98.3791 2.59109 109.429 2.59109C109.777 2.59109 108.741 2.72065 108.393 2.72065',
-				translateValues: [300, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['department'],
-		lines: [
-			{
-				path: 'M0.337217 5.52664C-2.52874 -0.498004 13.6269 4.31508 20.285 3.89095C30.6215 3.2325 63.9695 -3.00258 73.1254 1.80916C75.9565 3.29702 74.8157 68.2456 72.9132 79.1326C72.6533 80.6198 70.1049 93.8422 70.7911 94.0025C87.4182 97.8861 110.202 91.3884 127.664 91.0285C133.691 90.9043 139.648 90.4337 145.701 90.4337C147.39 90.4337 154.408 91.0298 153.765 90.8798',
-				translateValues: [280, 450]
-			}
-		]
-	},
-	{
-		letterGroupings: ['dv'],
-		specialOutlineMeanings: ['develop'],
-		lines: [
-			{
-				path: 'M0 5.89237C1.79185 4.009 21.5079 4.81136 24.7439 4.71688C41.9152 4.21554 86.2427 -4.72976 101.702 3.39446C112.147 8.88368 111.795 32.9159 120.364 40.4223C121.447 41.3709 141.531 6.3491 147.625 4.56995C149.816 3.93018 150.225 4.7611 148.254 5.45156',
-				translateValues: [230, 450]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['difference'],
-		lines: [
-			{
-				path: 'M0 0.69663C10.265 -0.502173 82.517 -0.73426 86.7941 4.8602C105.138 28.8538 70.4947 107.674 37.7735 105.381C13.8687 103.706 56.74 30.9194 63.8754 20.9197C67.4874 15.8576 83.3686 1.16386 81.9133 2.18362',
-				translateValues: [300, 460]
-			},
-			{
-				path: 'M38.8257 0C26.2058 0 -6.22777 0.457339 1.05222 23.0483C4.66302 34.2532 16.7343 38.7385 27.5785 35.539C29.9242 34.847 40.6043 30.4618 40.3112 31.0781',
-				translateValues: [400, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['different'],
-		lines: [
-			{
-				path: 'M0.37454 0C-2.41918 2.9364 11.3299 2.37235 11.4095 2.37918C38.4121 4.69605 65.8782 4.87942 92.8983 2.77571C93.7703 2.70782 102.955 1.33606 103.367 2.97398C111.588 35.6146 118.937 158.471 60.9253 152.664C8.9962 147.466 111.939 0.1813 101.953 6.54275',
-				translateValues: [210, 450]
-			},
-			{
-				path: 'M0 4.7603C13.8108 -1.69133 57.4265 3.25975 74.9811 2.38112C82.5733 2.00112 107.271 -0.0721639 106.954 0.00193803',
-				translateValues: [280, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['difficult'],
-		lines: [
-			{
-				path: 'M0 31.7473C0 33.2205 100.537 23.8159 103.385 32.4275C110.775 54.7753 78.3884 181.421 41.3927 161.95C10.5109 145.696 75.1309 43.2671 98.332 27.6664C109.982 19.8329 120.868 11.1073 132.923 4.15218C136.162 2.28385 146.484 -1.64212 143.611 0.751375C137.027 6.23821 106.04 14.9447 107.854 28.2494C109.238 38.394 146.418 31.8445 146.332 31.8445',
-				translateValues: [240, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['difficulty'],
-		lines: [
-			{
-				path: 'M0.00570769 27.658C-0.428172 28.114 24.0115 30.0372 26.532 30.0372C47.2258 30.0372 67.5659 27.269 88.2853 27.8067C90.3455 27.8602 93.2423 26.7239 94.4394 28.4015C97.2764 32.3774 99.1281 41.8596 98.4714 46.6915C95.8226 66.1799 64.8409 112.106 57.9392 105.279C27.8125 75.4762 110.868 0 139.428 0C143.583 0 131.756 3.19777 127.969 4.90706C119.83 8.58023 101.539 16.1414 100.806 27.9554C99.813 43.954 144.915 25.0242 147.704 27.9554C147.786 28.0412 133.523 52.8904 132.425 49.8141',
-				translateValues: [290, 430]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['discontinue'],
-		lines: [
-			{
-				path: 'M0.0250055 23.2311C-0.533882 22.0563 8.42259 23.4434 12.1917 23.6277C23.4535 24.1782 75.4148 29.7664 82.9286 17.4814C102.956 -15.2633 36.7896 5.86033 55.7656 17.6797C76.7732 30.7644 137.906 1.70294 155.646 22.8346C169.84 39.7425 144.72 71.8088 156.212 84.6934C166.038 95.7102 185.074 95.6941 191.297 81.5211C193.238 77.1016 196.315 62.8981 194.693 67.4443',
-				translateValues: [230, 430]
-			}
-		]
-	},
-	{
-		letterGroupings: ['dsc'],
-		specialOutlineMeanings: ['discount'],
-		lines: [
-			{
-				path: 'M0 17.5433C1.80647 19.442 3.67299 18.7796 7.07368 18.7329C15.3903 18.6186 130.664 30.8177 127.043 5.44908C124.865 -9.81294 57.9608 8.0955 70.4539 45.3004C75.3362 59.84 102.904 56.1749 114.028 52.8345C114.255 52.7662 128.15 48.3464 122.516 49.6622',
-				translateValues: [260, 440]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['e'],
-		lines: [
-			{
-				path: 'M0 1.58539e-05C0 -0.0161497 3.25077 12.3348 3.36842 15.8057C3.72459 26.3126 5.50441 37.0311 3.88664 47.5466C3.68471 48.8591 0.915179 58.7164 1.81377 59.2065C4.98583 60.9367 55.8191 59.4887 65.8138 60.1134C68.4376 60.2774 70.1017 59.7057 68.2753 62.4454',
-				translateValues: [290, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['e', 'ever', 'every'],
-		lines: [
-			{
-				path: 'M2.82947 0.0519852C2.82947 -0.981018 3.39537 13.6168 3.39537 22.8525C3.39537 34.6887 0 47.0291 0 58.342',
-				translateValues: [340, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['each'],
-		lines: [
-			{
-				path: 'M29.3967 0C29.3967 8.84317 33.1384 32.6249 27.699 40.2478C26.8638 41.4183 -10.8399 38.8733 3.08255 64.2379C3.91163 65.7484 4.69747 67.1056 6.19498 68.005C10.1259 70.3659 13.5686 70.5839 18.0788 70.979C21.6929 71.2955 12.9645 71.3755 21.4741 71.3755C27.9869 71.3755 32.1019 68.7239 37.8851 66.0223',
-				translateValues: [320, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['electric'],
-		lines: [
-			{
-				path: 'M5.6704 0C5.6704 5.56596 -0.295479 64.7487 0.0114494 65.8241C0.446732 67.3491 36.7902 61.4259 43.5853 61.8588C46.0957 62.0187 58.9312 63.137 56.8839 62.8501',
-				translateValues: [290, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['electrician'],
-		lines: [
-			{
-				path: 'M9.30437 0.0939696C10.3849 -0.716461 9.7582 3.97492 9.78943 4.21692C12.8768 28.1444 5.47071 41.0641 0.330902 62.9083C0.199974 63.4647 -0.16727 64.0947 0.0883754 64.606C1.07518 66.5796 4.43847 63.8287 6.63659 63.6359C24.1177 62.1024 41.5197 60.5854 59.0223 59.6342C62.7164 59.4334 78.7933 59.2123 75.029 57.3302',
-				translateValues: [270, 390]
-			},
-			{
-				path: 'M0 16.0602C5.17167 12.1815 6.38387 3.80704 13.0964 1.3874C44.7074 -10.0073 25.7078 52.1174 25.7078 66.1419',
-				translateValues: [360, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['electricity'],
-		lines: [
-			{
-				path: 'M4.2533 0C0.713023 1.65382 5.62836 24.3575 5.10214 29.8885C4.87662 32.2588 -0.244165 62.083 0.00909024 62.3049C2.55843 64.5378 57.5295 58.6187 70.0386 60.3718C85.6487 62.5594 63.6298 74.5693 56.6693 81.9331C51.8024 87.082 47.4309 91.2583 42.4512 95.9108',
-				translateValues: [330, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['electronic'],
-		lines: [
-			{
-				path: 'M2.26358 0C9.35483 7.45341 3.99548 45.2349 1.69768 56.5056C1.10589 59.4083 0 65.2293 0 65.2293C0 65.2293 38.8038 60.8674 58.2872 60.8674C61.6851 60.8674 68.4733 57.866 68.4733 61.264C68.4733 62.5428 43.7349 72.8419 41.8762 81.0905C38.6485 95.4146 55.6272 101.79 66.4926 94.176',
-				translateValues: [300, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['employ'],
-		lines: [
-			{
-				path: 'M9.19146 0C10.7602 1.64882 8.76014 10.0227 8.56238 11.7549C7.38685 22.0513 6.28668 31.7376 4.1588 41.8767C4.05083 42.3911 1.79866 60.2436 1.01339 60.2436C-2.96821 60.2436 5.88556 53.9342 8.56238 50.9867C16.1198 42.665 26.2513 34.5927 35.8226 28.6525C61.6423 12.6279 90.8604 6.47089 120.958 8.2284C149.128 9.87332 169.536 18.6337 191.206 35.2646C199.91 41.9443 219.295 58.6604 219.095 57.8927',
-				translateValues: [230, 390]
-			},
-			{
-				path: 'M115.519 0.236325C118.697 -1.24822 111.955 4.69664 110.486 5.96682C94.3643 19.9076 81.9426 37.2392 67.0795 52.3985C35.6968 84.4066 12.9809 126.694 2.91308 170.682C-18.5589 264.496 84.8496 280.253 110.906 230.044',
-				translateValues: [310, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['enclose'],
-		lines: [
-			{
-				path: 'M38.6214 0.060083C38.6214 -1.97547 31.5367 48.3872 36.9237 48.3872C42.6568 48.3872 25.2053 46.0775 19.7346 47.7924C9.10725 51.1239 -2.91745 69.8972 0.635695 81.101C7.67587 103.3 47.3927 83.8484 57.9325 78.5731',
-				translateValues: [330, 360]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['enclosed'],
-		lines: [
-			{
-				path: 'M35.7666 0C35.7666 15.1755 37.3598 32.2561 35.2007 47.3854C34.9101 49.4222 36.6922 53.5316 34.6348 53.5316C15.795 53.5316 -6.00989 61.2167 1.52999 87.6332C5.74479 102.4 55.29 96.478 55.29 86.0471',
-				translateValues: [330, 340]
-			},
-			{
-				path: 'M0 1.76493C0 -2.38549 8.30169 2.10498 12.4497 1.96319C24.612 1.54747 36.8149 2.77302 48.9499 1.96319C54.3176 1.60498 70.3165 2.592 65.0779 1.3684',
-				translateValues: [330, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['enclosure'],
-		lines: [
-			{
-				path: 'M25.1988 29.2977C27.6302 29.2977 29.6552 49.0538 29.6552 51.8999C29.6552 58.1363 33.9738 79.7188 29.6552 84.7624C25.4665 89.6542 -13.6776 87.6725 5.03877 116.287C16.1643 133.296 56.9894 121.363 55.5449 101.119C54.42 85.3549 15.6293 111.021 27.3209 115.84C46.52 123.754 117.409 48.5077 128.97 34.6508C130.938 32.2912 158.865 -0.414174 159.104 0.00397512',
-				translateValues: [310, 330]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['england'],
-		lines: [
-			{
-				path: 'M0 0C6.38205 13.416 -0.984512 62.2098 3.60758 65.4275C4.19562 65.8396 71.7913 59.5732 67.4829 60.966',
-				translateValues: [300, 390]
-			},
-			{
-				path: 'M0.342448 9.33455C-3.23692 9.33455 22.3001 2.81691 27.7176 1.15611C27.7909 1.13365 32.3862 -0.659103 32.3862 0.263914',
-				translateValues: [320, 470]
-			},
-			{
-				path: 'M0.0535157 12.9368C-1.29436 12.3072 23.2127 2.76254 29.1264 0',
-				translateValues: [320, 490]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['english'],
-		lines: [
-			{
-				path: 'M4.5184 0C4.5184 5.32581 -0.617797 59.9434 0.0619788 62.0075C0.593253 63.6206 3.45909 61.9502 5.15503 61.8588C22.5848 60.9193 39.6881 58.1582 57.1466 57.5465C63.1248 57.337 69.0127 56.7365 74.9723 56.3569C76.608 56.2527 81.1954 56.7026 79.8531 55.7621C74.9906 52.3549 47.3232 65.7736 50.3559 68.2528C51.0503 68.8205 51.8739 69.2193 52.6902 69.5911C61.3404 73.531 74.7684 66.6459 82.8241 73.9034C100.172 89.5321 62.3991 95.155 53.539 93.0855',
-				translateValues: [300, 390]
-			},
-			{
-				path: 'M0.314058 12.9368C-2.87625 12.9368 19.0191 5.43805 30.6602 0',
-				translateValues: [350, 500]
-			},
-			{
-				path: 'M0.186317 14.5725C-1.13302 14.5725 4.97636 12.4968 5.27937 12.342C13.1423 8.32453 20.6181 3.73919 28.6225 0',
-				translateValues: [360, 520]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['enthusiasm'],
-		lines: [
-			{
-				path: 'M8.00247 0C4.76382 4.85798 4.93498 20.4545 4.12205 26.4354C3.54344 30.6923 2.99377 34.9459 2.18184 39.168C-3.76517 70.0924 3.55957 38.0015 22.069 41.472C33.019 43.5251 37.9018 58.7381 37.3481 68.6349C36.8621 77.3234 28.368 190.328 28.1321 190.747C24.9174 196.462 16.3628 196.905 10.6703 196.446C7.87316 196.221 2.44766 195.039 1.21173 191.717C-16.4255 144.317 164.492 157.826 153.033 192.202',
-				translateValues: [250, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['enthusiastic'],
-		lines: [
-			{
-				path: 'M0.583112 0C0.583112 15.212 2.99441 28.3457 1.55477 43.8219C-0.777776 68.8968 -2.18633 51.5224 10.2997 45.2794C40.9306 29.9639 40.1956 83.2101 41.9758 100.567C43.6021 116.423 39.9033 135.508 38.8665 151.579C38.15 162.685 55.2463 233.568 21.1823 217.846C18.4797 216.599 12.2887 214.56 11.077 211.53C-1.09784 181.093 96.8869 185.76 117.571 183.935C122.735 183.48 133.118 183.547 133.118 183.547C133.118 183.547 129.325 183.184 127.482 183.547C119.769 185.062 86.5366 193.701 92.5021 208.615C98.5674 223.779 124.63 216.896 134.672 215.223',
-				translateValues: [300, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['equivalent'],
-		lines: [
-			{
-				path: 'M6.65166 0C2.31903 4.5539 6.65166 24.5075 6.65166 31.2973C6.65166 38.9899 -1.82735 51.1781 0.360833 58.3335C1.13773 60.8739 5.67715 58.6353 8.32921 58.4804C26.314 57.4302 44.2306 57.2706 62.2206 56.5703C62.6002 56.5555 64.565 55.9574 64.1079 56.5703C52.9968 71.4646 37.7481 82.4764 28.0405 98.4469C6.31602 134.187 -27.3587 241.34 41.4609 249.056C95.1678 255.077 120.344 161.217 128.694 121.516C132.702 102.456 136.061 83.8738 134.355 64.3578C134.18 62.3465 133.726 58.3335 133.726 58.3335C133.726 58.3335 144.513 78.6575 149.453 89.043C149.958 90.1029 154.183 102.414 155.535 102.414C155.716 102.414 175.968 72.5138 178.391 69.2067C180.121 66.8458 187.602 55.9028 187.408 56.7172',
-				translateValues: [280, 360]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['establish'],
-		lines: [
-			{
-				path: 'M17.4146 0C20.8276 0.797164 18.8352 6.90473 18.9001 10.4089C19.0469 18.3258 27.4389 79.7203 8.71402 73.1599C4.36213 71.6352 -0.281925 69.9155 0.013385 64.5353C1.28671 41.3372 74.0385 47.5837 88.9296 47.5837C93.0533 47.5837 118.433 47.8935 118.002 46.9889',
-				translateValues: [300, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['everything'],
-		lines: [
-			{
-				path: 'M0 0C0 15.4541 2.31989 31.1804 2.97095 46.6915C3.27768 53.9991 1.11263 73.7005 6.1541 68.4015',
-				translateValues: [300, 390]
-			},
-			{
-				path: 'M1.38183 67.7009C-3.47097 69.9679 5.90569 61.4929 7.32373 59.6712C22.6733 39.9525 40.6716 21.115 55.0711 0.935125C55.4165 0.451011 55.9325 -0.902094 55.4955 0.935125',
-				translateValues: [330, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['evidence', 'evident'],
-		lines: [
-			{
-				path: 'M0.679292 1.18166C-2.64826 1.18166 7.29167 9.71787 8.03592 10.5001C15.8979 18.7637 32.7409 51.9808 35.2797 51.9808C36.8994 51.9808 61.4192 17.4382 68.3037 2.96605C70.1399 -0.893951 72.3338 -0.650975 68.8696 1.77645',
-				translateValues: [280, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['examination'],
-		lines: [
-			{
-				path: 'M0.28534 0.628491C-0.106003 -0.194168 -0.240392 -0.224687 0.851235 0.628491C8.97687 6.97921 16.8694 13.3943 24.0529 20.8515C38.2478 35.5872 52.0373 51.4547 65.9291 66.056',
-				translateValues: [210, 390]
-			},
-			{
-				path: 'M0.0315072 62.696C-0.894128 63.6689 18.7587 46.5168 34.834 35.7319C60.1853 18.7238 85.5323 3.76 116.323 0.440711C141.783 -2.30395 174.189 7.94985 192.436 26.2152C202.134 35.9238 205.76 48.4569 213.091 58.7307C213.665 59.5352 216.12 62.3461 214.223 61.9029',
-				translateValues: [200, 390]
-			},
-			{
-				path: 'M0 22.1884C4.13086 17.8465 8.28915 14.6843 12.7326 10.8873C64.1696 -33.0672 36.8626 69.5714 43.2909 80.0818',
-				translateValues: [410, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['exchange'],
-		lines: [
-			{
-				path: 'M0.695884 0.432239C-2.61359 -1.88676 6.77954 5.76782 10.0331 8.16458C17.6052 13.7425 26.7674 17.5526 33.8007 23.778C35.7207 25.4774 40.9261 27.4669 36.9839 28.3876',
-				translateValues: [310, 250]
-			},
-			{
-				path: 'M60.4701 0.805096C33.7795 -5.42906 -3.05799 25.7325 0.202332 53.1471C1.68053 65.5767 32.5372 64.4312 40.0979 61.3256C43.9672 59.7362 51.7695 52.5329 51.7695 56.7159C51.7695 105.922 45.6154 155.1 45.6154 204.523',
-				translateValues: [310, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['exist'],
-		lines: [
-			{
-				path: 'M0 0C10.982 5.13015 23.6711 23.1284 33.5293 31.9703C34.8317 33.1385 41.9074 39.0602 41.1688 39.4052',
-				translateValues: [320, 280]
-			},
-			{
-				path: 'M1.13697 65.1315C-4.7901 71.3613 14.0427 53.7091 21.0848 48.7746C28.4669 43.6018 67.9873 14.355 58.0094 3.86753C46.2714 -8.46988 20.5333 12.1107 40.8203 18.2913C60.7395 24.3599 87.8007 16.841 108.303 16.5069C112.867 16.4326 140.028 16.792 133.556 17.5478',
-				translateValues: [310, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['expect'],
-		lines: [
-			{
-				path: 'M0 0C5.24353 11.0227 32.8571 30.3593 44.4227 38.4635C45.9255 39.5165 49.0022 40.2553 48.1011 41.834',
-				translateValues: [350, 270]
-			},
-			{
-				path: 'M126.499 0.739983C130.272 -3.22491 120.337 9.80776 117.728 14.6186C106.766 34.8315 103.084 58.2641 92.5457 78.6582C65.6039 130.799 50.3825 191.225 28.3166 245.796C19.0275 268.769 15.1816 301.732 2.56844 322.723C1.8976 323.839 -1.5978 327.085 0.870753 327.085',
-				translateValues: [270, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['experience'],
-		lines: [
-			{
-				path: 'M0 0C6.47343 3.02402 63.0701 16.6222 56.5895 18.4387',
-				translateValues: [270, 300]
-			},
-			{
-				path: 'M89.9773 0.935141C89.9773 -4.11483 87.4396 12.0132 80.3571 31.8645C75.8965 44.3669 70.501 56.0753 67.0585 68.9401C54.7722 114.856 41.77 159.217 24.0505 203.364C15.9775 223.477 5.6055 242.873 1.13179 264.033C0.660486 266.262 0 270.774 0 270.774C0 270.774 5.98156 264.171 9.33726 261.257C24.4049 248.175 37.3792 233.477 49.7987 217.837C81.6054 177.785 119.212 140.518 154.489 103.438C167.833 89.4134 185.746 81.8348 200.61 69.9315C206.465 65.2428 212.126 60.5031 217.304 55.0616C218.736 53.5561 224.192 51.0963 222.114 51.0963C213.087 51.0963 196.013 73.7325 190.141 78.2586C125.26 128.267 214.342 115.216 217.587 108.395',
-				translateValues: [210, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['extra'],
-		lines: [
-			{
-				path: 'M0.739053 0.994051C-2.65669 -4.09956 6.55931 11.7859 9.95505 16.8795C17.1525 27.6756 18.0461 39.322 23.7791 50.4694C25.6572 54.1214 31.613 60.2578 26.9319 59.3216',
-				translateValues: [220, 260]
-			},
-			{
-				path: 'M0 18.6745C64.8511 7.866 134.909 11.6421 200.569 8.12463C228.179 6.64552 255.759 5.5336 283.271 2.66779C285.47 2.43871 301.02 1.96986 298.065 0',
-				translateValues: [190, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['extraordinary'],
-		lines: [
-			{
-				path: 'M0.680796 0C-2.40691 0 5.97904 13.8774 6.51076 14.8664C10.9236 23.0743 17.782 41.6715 17.782 51.0121',
-				translateValues: [240, 250]
-			},
-			{
-				path: 'M0 10.9411C12.4045 15.076 32.0984 9.67147 45.4737 9.28932C101.094 7.70016 190.803 -5.77657 242.721 2.87637C243.173 2.95184 244.34 19.6433 252.049 21.9209C264.534 25.6098 284.502 17.042 284.502 2.29337',
-				translateValues: [190, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['f', 'from'],
-		lines: [
-			{
-				path: 'M2.61941 166.915C2.1895 166.915 2.06478 168.136 3.39674 166.656C12.6902 156.33 23.5085 144.022 30.7328 132.065C39.716 117.196 107.662 14.9276 71.672 1.08516C37.349 -12.116 11.7122 99.1004 6.3765 119.109C2.75061 132.707 2.48986 146.603 2.48986 160.437C2.48986 163.756 -1.71682 172.219 0.805647 167.174',
-				translateValues: [300, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['fact'],
-		lines: [
-			{
-				path: 'M84.5475 0.158963C84.5475 -1.15103 80.5512 5.99302 78.8178 8.4861C70.2899 20.7519 63.0279 37.1472 54.2014 48.4861C54.0857 48.6347 49.7945 48.0868 48.2595 48.04C39.2809 47.7665 -17.1275 56.7741 5.18077 74.3597C18.3854 84.769 123.732 81.028 144.815 80.159C157.993 79.6158 171.337 78.4461 184.499 78.0772C186.344 78.0255 199.75 77.7093 199.566 78.2259',
-				translateValues: [260, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['financial'],
-		lines: [
-			{
-				path: 'M55.223 16.3788C55.223 15.3214 55.223 18.4936 55.223 19.551C55.223 37.2919 62.4849 55.0888 62.8626 73.0827C63.3179 94.7801 65.685 241.237 11.6491 219.601C0.317943 215.064 2.36114 202.766 1.18004 192.835C-5.7378 134.666 17.7381 38.4047 72.1998 4.48289C104.985 -15.9379 115.643 38.8594 105.871 59.4024',
-				translateValues: [220, 390]
-			},
-			{
-				path: 'M53.7609 3.92847C60.9702 -3.64909 6.64471 -0.704191 3.11328 16.6174C-0.707142 35.3566 72.2202 17.0698 57.7221 47.5468C48.8472 66.2032 24.0269 57.0476 9.05518 56.2705C8.59242 56.2465 -4.59488 55.7813 1.69855 53.8913',
-				translateValues: [360, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['for ever and ever'],
-		lines: [
-			{
-				path: 'M110.525 70.5527C109.115 67.5898 94.7648 73.0319 92.6989 73.824C71.1621 82.0814 -38.5444 177.092 14.181 192.486C30.4129 197.225 110.526 76.4268 128.138 57.7645C146.048 38.7865 169.672 23.7341 188.83 5.57122C190.028 4.43548 193.971 -0.649884 193.287 0.0693636',
-				translateValues: [220, 260]
-			},
-			{
-				path: 'M1.64437 0C-2.17544 13.383 1.85658 29.8379 1.85658 43.8662',
-				translateValues: [440, 260]
-			},
-			{
-				path: 'M2.02668 0.00867439C-4.23048 -0.722072 6.5206 44.9566 0.965626 46.254',
-				translateValues: [460, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['for the'],
-		lines: [
-			{
-				path: 'M100.294 77.658C100.294 74.9533 85.9858 70.679 79.0733 73.0112C60.4953 79.279 38.9761 92.0493 26.8695 107.955C23.1847 112.797 -25.8593 190.68 18.3811 173.68C46.415 162.908 152.113 0 172.234 0C177.527 0 169.731 10.288 168.626 15.4647C164.248 35.9827 166.097 57.5025 162.684 78.0669C160.233 92.8398 161.971 181.96 157.167 183.643',
-				translateValues: [260, 280]
-			}
-		]
-	},
-	{
-		letterGroupings: ['frt'],
-		specialOutlineMeanings: ['fortunate'],
-		lines: [
-			{
-				path: 'M112.219 48.2061C92.1333 52.8976 76.5825 56.5065 59.8031 69.47C46.4725 79.769 -33.8426 169.561 16.2999 180.102C30.7602 183.142 99.884 72.6648 117.1 53.708C132.054 37.2416 145.024 22.1593 154.025 1.66334C155.011 -0.583016 155.088 -0.425528 154.449 1.36594',
-				translateValues: [260, 280]
-			},
-			{
-				path: 'M0 4.68628C34.686 4.68628 68.2789 5.05057 102.71 0.225309C107.721 -0.47696 103.951 0.629358 101.224 1.2662',
-				translateValues: [370, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['from the'],
-		lines: [
-			{
-				path: 'M54.3259 139.821C53.4183 137.913 58.2894 136.151 58.5701 135.955C72.8908 125.92 86.7184 114.087 100.376 103.092C112.466 93.3583 178.588 12.6286 147.699 1.08475C94.4414 -18.8182 0 241.282 0 279.003',
-				translateValues: [290, 320]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['from time to time'],
-		lines: [
-			{
-				path: 'M88.2904 2.89931C88.9929 3.63774 88.8563 3.22044 88.8563 4.0889C88.8563 34.8162 57.9333 252.835 11.3287 209.293C-30.2093 170.485 53.4618 38.5137 85.7439 6.46808C97.5219 -5.22374 145.514 2.86412 162.988 1.11492C165.626 0.850922 173.828 -1.64653 172.043 2.10624',
-				translateValues: [210, 310]
-			},
-			{
-				path: 'M0.170867 7.13755C-1.2877 10.2037 6.97605 7.45044 10.357 7.13755C37.7965 4.59809 68.2902 0 95.8071 0',
-				translateValues: [340, 280]
-			},
-			{
-				path: 'M0 6.34449C31.7801 2.63303 66.878 0 99.0316 0',
-				translateValues: [390, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['g', 'go', 'gentleman'],
-		lines: [
-			{
-				path: 'M3.17134 22.6047C-2.12471 22.1634 28.0044 -1.36271 40.3535 0.0621974C65.0871 2.91607 71.4982 18.5035 72.3535 40.7424C73.7586 77.2741 42.8407 115.096 25.0661 144.386C18.0119 156.01 11.9221 168.059 5.24421 179.884C4.39668 181.385 -0.611003 194.102 0.0620246 186.362',
-				translateValues: [310, 390]
-			}
-		]
-	},
-	{
-		letterGroupings: ['gn'],
-		specialOutlineMeanings: ['general'],
-		lines: [
-			{
-				path: 'M0.150621 26.5738C-2.54013 26.5738 31.6423 6.71442 37.0146 5.11019C63.4769 -2.79175 103.633 -6.2443 105.65 31.0605C108.534 84.4162 56.9947 219.724 8.39652 252.123C4.78045 254.534 13.9519 245.061 17.855 243.15C26.9903 238.676 44.7352 228.488 55.6891 235.631C65.8332 242.247 66.9771 289.496 67.8155 299.416C68.5086 307.618 64.6139 330.46 69.7557 324.032',
-				translateValues: [270, 310]
-			}
-		]
-	},
-	{
-		letterGroupings: ['gv'],
-		specialOutlineMeanings: ['government'],
-		lines: [
-			{
-				path: 'M0.765514 17.3209C-3.25491 20.8424 9.77164 11.0124 11.8005 9.78683C22.8142 3.13383 35.1769 -0.57086 48.0177 0.0718349C68.2964 1.0868 73.5357 33.1527 76.3125 47.259C90.8108 120.913 2.74615 201.478 2.74615 225.499C2.74615 233.696 14.8483 248.571 20.0059 255.437C21.7181 257.717 25.3888 272.293 27.0796 273.083C27.1968 273.138 34.4663 262.516 35.2851 261.187C42.7342 249.099 52.8056 238.705 59.6186 226.292C60.1592 225.308 62.9433 219.552 63.5798 221.336',
-				translateValues: [290, 320]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['guilty'],
-		lines: [
-			{
-				path: 'M26.6852 11.1263C24.8283 11.1263 25.1081 10.8399 27.1096 9.78803C49.6975 -2.08273 85.2881 -9.58035 87.1652 26.591C88.2556 47.6026 5.33624 166.219 1.21992 166.219C-3.92209 166.219 8.61361 159.066 12.4671 155.662C23.5226 145.894 32.9087 139.727 46.8452 134.844C48.7575 134.174 55.2664 130.618 54.9092 131.87',
-				translateValues: [310, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['h', 'he'],
-		lines: [
-			{
-				path: 'M2.46154 0C8.0645 16.8089 0 158.953 0 196.016',
-				translateValues: [330, 260]
-			}
-		]
-	},
-	{
-		letterGroupings: ['hf'],
-		specialOutlineMeanings: ['half'],
-		lines: [
-			{
-				path: 'M1.48547 0.146364C1.48547 -1.58454 1.89106 12.4238 1.48547 21.7077C0.870813 35.7773 1.06105 50.1174 1.06105 64.2356C1.06105 82.332 3.20999 207.777 9.54947 210.853C50.196 230.571 26.4581 81.4015 0 87.5813',
-				translateValues: [340, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['hand'],
-		lines: [
-			{
-				path: 'M0 0C3.14608 0 124.913 111.923 140.156 124.839C155.12 137.52 170.081 152.03 182.554 167.325C186.96 172.728 201.814 191.288 199.908 189.953',
-				translateValues: [260, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['hospital'],
-		lines: [
-			{
-				path: 'M0.512021 0C1.69335 0.885996 0.987694 0.211686 1.28935 3.30364C6.35653 55.2422 1.14155 105.971 4.01 157.603C4.61856 168.557 -0.883166 178.693 0.123357 189.765C1.51503 205.074 33.3651 188.979 18.5849 181.02C16.6867 179.998 11.3573 177.373 9.06263 178.98C-4.28766 188.325 4.05228 214.49 5.75898 227.077C9.62754 255.607 6.08105 293.166 2.844 322.3C1.52873 334.137 1.79657 345.761 0.123357 357.474C-0.230368 359.95 1.3864 368.818 1.87234 367.968',
-				translateValues: [350, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['however'],
-		lines: [
-			{
-				path: 'M6.36632 0C6.36632 47.4927 11.8323 98.1928 4.88084 145.279C1.4313 168.645 0 192.044 0 215.911C0 216.464 0 219.467 0 217.398',
-				translateValues: [330, 240]
-			},
-			{
-				path: 'M1.90989 0C-0.987056 0 2.98964 34.8934 2.12211 39.5539C1.78329 41.3741 1.19543 52.3057 0 51.7472',
-				translateValues: [360, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['husband'],
-		lines: [
-			{
-				path: 'M44.192 0C44.192 49.7606 32.1234 101.802 34.6425 151.227C36.1892 181.571 71.3682 152.811 63.5032 145.725C22.3702 108.668 -3.9105 223.041 0.476654 246.097C15.3841 324.44 111.992 307.889 99.1545 243.123C90.4047 198.978 6.20634 218.388 6.20634 261.115',
-				translateValues: [310, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['i', 'I', 'eye'],
-		lines: [
-			{
-				path: 'M28.4331 0C30.553 1.4133 25.1201 6.78327 24.5465 7.64372C17.4059 18.3546 10.2725 28.9518 2.91084 39.5142C-10.7314 59.0879 27.8155 24.2879 31.2833 22.1538C40.5342 16.461 49.6972 11.2542 57.4533 3.49798',
-				translateValues: [350, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['I think'],
-		lines: [
-			{
-				path: 'M0 39.0249C0 43.7074 32.1392 -1.96425 33.1048 0.0657699',
-				translateValues: [310, 420]
-			},
-			{
-				path: 'M0.752855 36.4503C-5.58087 37.0843 30.1401 -0.972446 29.8257 0.0190436',
-				translateValues: [340, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['identification'],
-		lines: [
-			{
-				path: 'M29.6834 0.0204533C27.7177 -0.897799 10.3686 29.3802 7.33054 34.9151C5.37162 38.484 -2.96545 45.4232 1.1057 45.4232C24.578 45.4232 49.1199 44.8284 72.9743 44.8284C77.3159 44.8284 101.448 45.349 94.7613 44.8284',
-				translateValues: [280, 410]
-			},
-			{
-				path: 'M0 12.3673C8.9795 -1.65153 24.5234 -3.54087 34.8025 5.62624C53.2712 22.0969 41.8762 56.2844 41.8762 77.3983',
-				translateValues: [370, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['if the'],
-		lines: [
-			{
-				path: 'M100.057 0C87.0115 27.4237 93.9249 61.024 81.604 88.2034C73.2984 106.525 65.4712 125.707 56.1212 143.465C13.8074 223.832 -22.7549 221.381 17.2379 129.303C30.1331 99.6143 49.4446 68.7528 72.1578 45.4101C76.1763 41.2802 97.4154 42.4008 103.792 41.4079C120.788 38.7613 138.282 38.1239 155.416 36.7899C156 36.7444 156.834 36.313 157.174 36.7899C164.001 46.3583 162.609 135.562 163.105 153.625C163.467 166.825 165.082 180.012 165.082 193.185C165.082 199.749 169.256 218.99 169.256 212.427',
-				translateValues: [260, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['immediate'],
-		lines: [
-			{
-				path: 'M34.7068 0.110818C37.418 -1.31403 29.3128 11.4453 28.3404 12.8989C26.0808 16.2768 3.61739 51.1146 0.540862 51.1146C-1.90845 51.1146 4.63704 48.4279 6.69497 47.0997C13.9698 42.4045 20.7229 36.8967 27.916 32.2298C47.0649 19.8058 69.5287 11.7021 92.6402 10.6684C134.738 8.78554 144.235 27.0219 166.065 53.7911C167.764 55.8749 184.193 62.4441 224.423 53.0476',
-				translateValues: [250, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['immediately'],
-		lines: [
-			{
-				path: 'M27.1399 4.35933C31.3367 8.77042 4.41127 49.6103 0.189205 58.4857C-1.10343 61.2031 4.5639 54.3315 6.97994 52.5378C15.701 46.0631 24.1009 39.0023 32.8696 32.6121C55.5835 16.0595 84.6815 0.912404 113.085 0.0470616C142.853 -0.859846 181.331 11.3286 198.182 37.6679C200.876 41.88 202.134 54.1176 206.033 56.8501C216.855 64.4326 286.408 49.3184 305.136 51.943C306.092 52.077 300.693 58.6177 298.982 60.7162C292.203 69.0287 283.105 85.9679 274.365 92.0917',
-				translateValues: [220, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['importance'],
-		lines: [
-			{
-				path: 'M38.216 3.86817C26.741 15.9293 13.8227 27.9831 5.67707 42.7282C5.00102 43.9519 0.440633 54.8223 0.0181271 54.8223C-0.352201 54.8223 5.05797 51.2908 6.52592 50.2622C21.6639 39.6548 37.4019 30.3711 53.4952 21.3155C79.519 6.67202 153.031 -14.7495 176.011 14.7728C198.209 43.2885 189.001 139.594 180.539 176.161C178.929 183.118 170.48 223.699 171.484 234.252C171.941 239.056 170.085 253.42 171.201 248.726',
-				translateValues: [230, 400]
-			},
-			{
-				path: 'M42.9721 2.6517C42.9721 -3.79756 22.1575 3.38705 19.4875 4.63435C-18.9539 22.592 6.11742 65.7502 37.5961 43.6926',
-				translateValues: [440, 590]
-			}
-		]
-	},
-	{
-		letterGroupings: ['imp'],
-		specialOutlineMeanings: ['important'],
-		lines: [
-			{
-				path: 'M31.6328 0C32.12 1.02423 31.8825 0.163041 31.6328 1.38786C27.9727 19.3407 11.5986 39.3923 4.18692 56.7039C4.06859 56.9803 1.81512 63.4449 1.0745 63.4449C-3.05311 63.4449 5.82 56.5632 8.99703 53.9281C21.4059 43.636 36.6253 34.5244 50.5903 26.171C88.6116 3.42776 149.34 -14.4002 188.386 20.4213C228.762 56.4296 191.779 205.786 201.401 259.727',
-				translateValues: [240, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['impossible'],
-		lines: [
-			{
-				path: 'M36.2362 1.02521C33.9107 -3.86327 30.2949 10.0785 27.1819 14.5072C19.2843 25.7428 13.4803 37.3789 7.94148 49.7985C2.70205 61.5468 0.133996 67.1282 0.0189516 67.0475C-1.01211 66.3251 40.354 37.5347 44.4417 35.3251C78.0037 17.1831 123.876 -9.25594 157.621 22.8344C190.844 54.4284 175.176 94.803 177.71 135.647C178.567 149.465 175.139 163.544 173.829 177.763C171.951 198.168 161.558 238.86 172.051 257.977C184.377 280.434 217.828 238.015 191.291 233.789C185.214 232.821 176.982 229.456 172.617 235.573',
-				translateValues: [230, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['improve'],
-		lines: [
-			{
-				path: 'M49.5158 0C34.02 16.2871 21.8161 40.3125 8.20547 58.8848C5.26081 62.9029 0 71.3755 0 71.3755C0 71.3755 9.91316 62.564 15.2792 58.6865C39.2985 41.3298 63.5331 24.5665 92.2408 16.2578C159.595 -3.23617 188.641 26.3795 229.187 68.9963',
-				translateValues: [220, 390]
-			},
-			{
-				path: 'M0 0C11.0162 13.0742 56.3622 78.7113 56.8724 78.7113C61.655 78.7113 106.177 16.1881 115.19 5.75414',
-				translateValues: [310, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['individual'],
-		lines: [
-			{
-				path: 'M48.1011 0C45.4745 0 37.3277 10.0625 35.0855 12.4907C30.4511 17.5094 0 42.4378 0 51.1524C0 55.0516 88.692 48.7404 96.576 49.0473C104.345 49.3497 120.755 87.442 127.326 93.5812C127.762 93.9878 150.801 54.1188 157.056 51.9273C175.958 45.3049 192.292 51.3736 212.211 49.1698C218.495 48.4745 241.345 47.4276 237.959 48.3767',
-				translateValues: [220, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['inform'],
-		lines: [
-			{
-				path: 'M39.012 0C31.5492 15.6879 15.2575 36.6007 2.22883 48.7733C-0.633801 51.4478 7.98989 43.2352 11.5661 41.6357C28.7736 33.9394 43.0955 27.6461 51.4617 47.1871C59.0428 64.8945 53.4746 226.65 43.8221 243.866C39.4634 251.641 32.896 257.139 25.4305 261.908C20.2315 265.23 12.5583 269.462 7.03893 264.684C-19.325 241.864 35.6818 129.376 52.0276 106.468',
-				translateValues: [290, 340]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['information'],
-		lines: [
-			{
-				path: 'M30.4346 0C26.8504 7.53451 -3.53059 44.0216 1.57398 41.0409C7.56007 37.5455 17.3801 34.1679 24.4927 33.3086C40.2908 31.3999 41.9878 49.1157 43.1672 60.2726C45.638 83.6455 39.1753 233.912 14.5896 232.763C-17.0574 231.285 11.4554 153.867 19.3997 141.561C24.0708 134.326 29.2394 127.215 34.6788 120.545C35.5498 119.477 42.361 109.978 40.3378 112.813',
-				translateValues: [290, 300]
-			},
-			{
-				path: 'M0 14.7899C3.839 1.67171 15.981 -3.53635 26.3141 2.49744C38.1857 9.42958 34.9682 63.2902 31.9731 77.2434',
-				translateValues: [390, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['insurance'],
-		lines: [
-			{
-				path: 'M36.9857 0C25.4621 12.1121 7.55894 31.7229 0.202539 47.1871C-1.19664 50.1284 5.03216 43.7657 7.27622 42.627C18.7475 36.8064 34.3722 32.1995 39.5322 47.3854C44.0641 60.7228 26.3596 113.734 36.4198 115.589C53.3821 118.716 60.8631 82.0724 42.6446 88.0298C41.4021 88.4361 35.0051 91.4799 35.0051 91.9951',
-				translateValues: [300, 340]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['intention'],
-		lines: [
-			{
-				path: 'M34.771 0.00998729C31.1151 -0.843924 2.16491 53.3433 0.534341 53.3433C-1.60756 53.3433 3.26215 49.9452 5.0615 48.7832C10.9302 44.9935 28.7011 30.0756 34.488 42.2405C47.5045 69.603 38.7322 99.2368 38.7322 128.684',
-				translateValues: [310, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['involve'],
-		lines: [
-			{
-				path: 'M42.5376 0C42.5376 8.32824 22.9247 27.0074 17.0723 34.6964C14.3723 38.2438 3.89029 62.057 0.0955022 62.057C-1.34366 62.057 13.7465 42.1308 27.5414 48.575C47.0027 57.6662 51.6156 80.0057 61.4951 96.1586C63.4534 99.3606 66.0222 106.468 66.0222 106.468C66.0222 106.468 71.0176 97.17 73.9448 92.7881C87.8056 72.0387 104.119 59.6968 120.065 42.4288C122.078 40.2489 122.329 39.7482 122.329 42.0322',
-				translateValues: [260, 350]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['is the'],
-		lines: [
-			{
-				path: 'M80.7488 0C85.1757 1.10671 69.0072 14.1365 66.9512 15.8381C52.6056 27.7103 35.9605 53.5385 15.2589 53.5385C10.8514 53.5385 -1.2244 52.3659 0.101027 45.4737C0.364032 44.1061 2.64677 40.8875 3.40467 40.1296C54.4819 -10.9477 29.9043 165.451 28.8622 181.603C28.0723 193.846 27.7894 206.16 27.1132 218.332C26.7541 224.796 27.2134 229.892 28.0848 236.211C28.3819 238.364 28.4735 244.895 28.4735 242.721',
-				translateValues: [320, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['it is my pleasure'],
-		lines: [
-			{
-				path: 'M29.995 3.92666C33.5243 3.92666 6.58005 39.8134 0.0732873 44.3728C-0.334252 44.6583 1.06859 44.3099 1.55876 44.2241C14.206 42.0085 27.0118 39.7223 39.5444 36.9378C63.5833 31.5969 91.78 26.9504 116.365 29.3542C163.98 34.0097 107.027 73.4719 107.027 44.5215C107.027 40.4418 106.897 36.921 108.937 33.2204C111.489 28.5928 115.037 24.1044 119.123 20.7296C132.345 9.8108 147.83 6.28152 164.324 2.73707C236.935 -12.8659 264.225 43.3319 272.339 43.3319C276.647 43.3319 293.045 22.0911 297.805 17.7557C299.639 16.085 300.079 16.2546 298.229 15.8226',
-				translateValues: [200, 270]
-			},
-			{
-				path: 'M87.9754 0C72.2101 3.68234 54.8747 38.9303 45.1089 50.4089C6.57875 95.697 -30.1048 172.056 38.106 209.963C54.8735 219.281 95.6042 223.084 92.6441 191.97C89.6854 160.872 31.6566 227.459 75.455 210.409C99.2333 201.152 117.153 164.635 132.115 145.874C139.302 136.862 146.212 127.713 153.548 118.81C155.27 116.721 164.66 107.78 162.461 108.55',
-				translateValues: [360, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['j'],
-		lines: [
-			{
-				path: 'M107.919 0.642606C107.919 -2.51698 106.65 6.83834 106.235 9.97054C104.807 20.7439 106.217 31.8003 105.717 42.6183C104.403 70.9912 103.825 100.165 100.405 128.383C96.7319 158.685 89.3711 208.25 65.9433 231.12C52.6349 244.112 36.5294 252.595 19.5628 259.234C16.7182 260.347 0.504738 258.122 0 260.141',
-				translateValues: [290, 330]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['job'],
-		lines: [
-			{
-				path: 'M43.3547 0.0434499C42.3883 -0.972265 50.8801 16.1223 52.4797 19.6717C58.8525 33.8126 64.2446 48.5884 68.1833 63.5379C78.1858 101.503 89.5187 151.765 48.0233 172.534C39.0736 177.014 26.7255 181.076 16.6161 178.333C-17.5712 169.06 8.80225 130.999 26.3778 121.828C40.1582 114.637 58.0438 110.974 73.4886 114.839C74.7027 115.143 87.6495 117.481 79.4305 117.07',
-				translateValues: [310, 350]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['jury'],
-		lines: [
-			{
-				path: 'M68.2668 0.54815C68.2668 -2.178 68.0275 6.01096 68.2668 8.72659C69.1682 18.9599 71.6363 28.944 72.511 39.2099C74.9433 67.7578 70.4745 94.4026 65.7202 122.481C58.3125 166.232 27.9221 195.122 5.24023 230.883C-7.06444 250.282 4.61898 234.842 16.4874 225.232C26.8794 216.818 45.5772 215.414 58.2929 212.444',
-				translateValues: [280, 340]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['k', 'kind', 'like', 'knowledge'],
-		lines: [
-			{
-				path: 'M43.179 0.684699C44.7946 -2.00801 40.6329 4.03878 39.4219 5.08956C32.659 10.9574 1.39402 31.3493 0.0372553 40.8466C-0.659749 45.7257 8.60028 45.7367 12.9927 47.9721C23.4798 53.3093 33.0897 59.9886 43.3085 65.7211C45.1999 66.7822 54.3741 68.1827 49.0089 68.1827',
-				translateValues: [340, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['know ', 'and ', 'new'],
-		lines: [
-			{
-				path: 'M0.159054 13.8725C-1.41044 13.3493 9.14418 4.28474 9.48699 4.02631C21.5392 -5.05921 28.8456 2.5343 34.3615 13.8725C40.4194 26.3249 36.8725 47.172 36.3048 60.5121C35.801 72.3504 36.8564 83.0587 36.823 82.9251',
-				translateValues: [340, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['l', 'letter ', 'local'],
-		lines: [
-			{
-				path: 'M69.2485 0.327705C72.1617 -3.70596 42.8325 30.6871 37.2485 38.1577C16.2587 66.2385 1.71265 105.214 0.066286 140.117C-1.67057 176.939 30.9692 224.8 69.2485 193.882C70.5925 192.797 83.7117 184.849 81.5562 180.538',
-				translateValues: [310, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['ladies and gentlemen'],
-		lines: [
-			{
-				path: 'M72.898 0C39.1048 11.2644 -6.8886 106.317 0.865583 139.66C5.84576 161.075 29.8074 167.407 45.3028 152.486C47.3425 150.522 57.6439 135.773 56.5741 135.773',
-				translateValues: [330, 290]
-			},
-			{
-				path: 'M6.22052 13.4261C6.28149 13.609 6.3323 14.1933 6.73874 13.6852C13.5625 5.15557 21.6861 -1.26291 33.2974 0.211551C51.8997 2.57374 68.9044 14.5389 66.8521 35.0618C64.0536 63.0467 49.1985 96.8419 29.7995 117.588C22.7525 125.124 15.8893 132.767 8.68206 140.131C6.95546 141.895 -0.133047 150.128 0.00190121 149.588',
-				translateValues: [360, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['landlord'],
-		lines: [
-			{
-				path: 'M52.7262 0C48.9609 0 38.5389 10.5545 36.1738 12.6394C-9.95853 53.3071 -14.9907 153.676 38.7203 103.494',
-				translateValues: [280, 340]
-			},
-			{
-				path: 'M93.8351 0C93.8351 5.63336 75.7561 17.9966 72.1896 21.8587C48.5675 47.4396 24.1152 77.1026 8.31427 108.104C-7.37923 138.895 -3.76094 206.454 44.6023 200.149C59.654 198.187 76.239 177.119 78.1315 162.974C78.2258 162.269 78.2124 158.098 78.3437 158.067C94.6439 154.26 112.859 154.944 129.699 154.944',
-				translateValues: [350, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['m', 'me', 'my'],
-		lines: [
-			{
-				path: 'M0.382651 30.7188C-3.7565 30.7188 26.9068 13.4188 33.4191 11.156C58.7815 2.34367 85.9006 -3.37205 112.447 2.21675C116.657 3.10299 155.888 25.2944 151.443 28.2572',
-				translateValues: [250, 430]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['man and woman', 'men and women'],
-		lines: [
-			{
-				path: 'M0.0234719 43.9248C-0.789891 45.9196 19.7608 27.7701 26.6546 24.5293C52.9972 12.1457 75.5017 1.82478 104.871 0.137965C127.788 -1.17829 150.242 6.99172 167.779 21.7375C174.029 26.9928 180.14 32.9405 185.393 39.0759C186.396 40.2474 192.504 44.6551 190.006 44.0717',
-				translateValues: [210, 410]
-			},
-			{
-				path: 'M0 0.000755005C0.926303 -0.215603 72.5016 46.1438 113.654 42.0244C131.286 40.2594 148.858 31.8253 164.4 23.6574C171.085 20.1443 186.412 4.2661 186.208 4.40883',
-				translateValues: [320, 450]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['man or woman', 'men or women'],
-		lines: [
-			{
-				path: 'M1.52514 184.154C-4.82183 184.154 10.3768 175.002 15.365 171.077C31.9405 158.036 46.4315 147.528 67.5788 143.453C116.544 134.018 143.486 181.803 156.279 181.803C171.15 181.803 342.52 0.484649 355.489 0.484649C369.317 0.484649 387.809 39.6826 420.704 36.0431C443.376 33.5347 466.784 21.5985 482.564 5.6274C483.462 4.71826 490.405 -0.530866 490.951 0.0438423',
-				translateValues: [110, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['manufacture'],
-		lines: [
-			{
-				path: 'M0 107.955C3.90158 108.867 39.4532 76.499 57.0846 73.1599C82.2456 68.3948 111.735 76.7665 130.509 93.6803C134.453 97.2332 136.136 107.437 139.635 109.889C143.483 112.585 145.704 102.54 149.608 99.9257C157.483 94.6531 177.239 84.9032 186.321 93.0855C202.448 107.614 210.706 243.435 180.591 245.353C151.086 247.233 160.428 188.922 166.161 175.762C190.037 120.956 233.672 43.4649 275.025 0',
-				translateValues: [240, 350]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['maximum'],
-		lines: [
-			{
-				path: 'M0.756366 37.7888C-5.6188 39.7034 30.344 19.5467 32.3757 18.458C56.0493 5.77273 78.1449 -0.392258 105.164 0.0193199C121.774 0.272341 141.045 8.36312 153.124 19.6476C158.236 24.4237 163.273 29.2954 167.766 34.6662C168.565 35.6212 171.898 40.6683 171.586 39.5732',
-				translateValues: [260, 420]
-			},
-			{
-				path: 'M2.49198 40.8922C-4.13376 42.4398 4.32747 36.9567 6.31177 35.539C24.0015 22.9005 41.128 10.6104 60.6377 1.04089C61.1522 0.788513 63.3509 0.336327 61.9109 0',
-				translateValues: [390, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['mayor'],
-		lines: [
-			{
-				path: 'M0.727505 107.509C-2.14925 108.181 4.33353 103.976 5.82056 102.9C12.9241 97.7562 19.3397 92.9824 27.2538 89.2194C45.2952 80.6409 70.6714 70.2749 90.917 76.58C104.84 80.916 129.598 108.104 135.481 108.104C139.248 108.104 237.11 10.4229 251.985 0',
-				translateValues: [240, 350]
-			}
-		]
-	},
-	{
-		letterGroupings: ['mm'],
-		specialOutlineMeanings: ['member'],
-		lines: [
-			{
-				path: 'M0.386772 37.9313C-6.39166 30.8067 77.8593 -0.331569 101.611 3.13578C119.09 5.68744 139.112 13.5791 153.391 24.251C155.767 26.0269 166.101 36.5961 168.245 36.2956C203.296 31.3835 214.237 -3.64358 268.409 0.310497C285.449 1.55427 308.818 9.25974 323.796 18.0057C330.76 22.0723 336.424 25.0226 342.682 30.3477C344.305 31.728 350.726 37.7106 350.534 37.0391',
-				translateValues: [190, 420]
-			}
-		]
-	},
-	{
-		letterGroupings: ['mn'],
-		specialOutlineMeanings: ['minimum'],
-		lines: [
-			{
-				path: 'M0.804734 49.9807C-6.35725 53.0502 36.3514 24.1112 47.1273 18.5736C81.695 0.809579 127.772 -5.36108 165.48 5.11337C189.607 11.8154 208.933 31.3386 223.201 51.3146C223.628 51.9113 225.33 55.4376 226.112 55.4376C230.527 55.4376 250.31 23.7437 260.793 42.4624C265.77 51.3499 275.146 102.766 276.315 113.28C276.525 115.175 276.965 121.724 278.74 119.95',
-				translateValues: [210, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['motorway'],
-		lines: [
-			{
-				path: 'M3.48069 60.1375C-6.28037 63.5574 7.43938 55.7884 8.85669 54.9826C28.6078 43.7541 48.9927 29.8324 70.8222 23.0619C130.297 4.6157 191.746 -7.42456 253.606 5.21806C285.53 11.7426 319.917 24.6635 348.676 39.3197C356.845 43.4826 386.468 63.9604 384.611 63.3098',
-				translateValues: [130, 240]
-			},
-			{
-				path: 'M0 3.44141C1.99695 2.97498 50.6341 41.6776 76.1128 38.9309C94.0071 37.0018 107.827 27.0665 120.819 15.5356C124.59 12.1881 128.249 8.98966 131.853 5.62233C133.298 4.27283 138.76 -0.645604 138.078 0.070901',
-				translateValues: [450, 310]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['much'],
-		lines: [
-			{
-				path: 'M3.34206 33.5086C-11.1854 42.3865 26.0698 18.6155 30.7475 16.7743C56.8207 6.51146 90.064 -3.20887 118.542 1.01009C136.79 3.71355 161.257 14.8336 172.625 29.992C176.469 35.1163 128.502 24.605 130.911 53.517C132.474 72.2741 171.88 58.2195 171.413 57.5187',
-				translateValues: [240, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['n', 'and'],
-		lines: [
-			{
-				path: 'M0.159054 13.8725C-1.41044 13.3493 9.14418 4.28474 9.48699 4.02631C21.5392 -5.05921 28.8456 2.5343 34.3615 13.8725C40.4194 26.3249 36.8725 47.172 36.3048 60.5121C35.801 72.3504 36.8564 83.0587 36.823 82.9251',
-				translateValues: [340, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['necessary'],
-		lines: [
-			{
-				path: 'M0.0329228 20.2286C-0.927077 8.16026 19.3497 -14.452 29.1765 13.0911C30.826 17.7145 18.6648 103.783 40.7773 89.0267C75.9797 65.5345 19.071 44.4182 3.14534 122.534',
-				translateValues: [300, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['nevertheless'],
-		lines: [
-			{
-				path: 'M1.07806 119.476C-4.61628 122.136 13.099 99.177 34.1829 120.517C48.9702 135.484 53.9711 159.085 67.4999 173.305C69.2351 175.129 69.5051 179.974 71.9564 179.402C78.7467 177.816 104.071 141.022 109.093 135.089C146.339 91.0951 200.011 30.9595 249.364 0.219513C251.817 -1.30816 242.379 5.36575 231.963 13.1563C223.201 19.7102 214.239 26.875 205.861 33.9742C181.88 54.2956 148.623 88.0776 140.288 119.03C118.794 198.853 216.044 173.841 205.649 139.848C199.89 121.015 159.205 148.921 178.062 160.666C182.114 163.19 191.334 166.154 194.614 161.558',
-				translateValues: [250, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['newspaper'],
-		lines: [
-			{
-				path: 'M0.570396 17.333C-4.31293 17.333 23.5514 -7.31718 31.1453 2.17514C50.0137 25.7606 20.4257 62.9948 34.3842 87.4221C42.3662 101.391 60.2259 80.2818 51.0967 70.4504C25.7319 43.1345 36.31 186.813 38.1412 193.527C39.3352 197.905 35.6823 205.9 40.0846 207.001C42.5838 207.626 49.3272 186.886 49.9307 184.977C58.6975 157.245 70.5687 127.389 88.0198 103.746C96.5958 92.1269 107.017 82.0938 114.449 69.8027C116.571 66.2929 126.333 53.9703 124.425 56.0699',
-				translateValues: [300, 360]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['next'],
-		lines: [
-			{
-				path: 'M0 19.2323C40.7458 -23.5944 59.0385 12.6255 68.1196 54.6227C70.4628 65.4595 75.9624 71.7041 80.4278 81.0911C81.1651 82.641 81.5178 83.1661 80.2156 83.4703',
-				translateValues: [280, 370]
-			},
-			{
-				path: 'M0 9.47643C0 5.44413 8.07085 9.56778 12.096 9.32774C27.3057 8.4207 43.2949 6.37267 58.3579 4.12327C61.2446 3.69219 82.5321 -0.743505 77.669 0.108402',
-				translateValues: [310, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['no doubt'],
-		lines: [
-			{
-				path: 'M0.00902935 19.5165C0.00902935 19.9641 -0.125695 20.3921 0.645661 19.6652C6.83838 13.8296 25.7661 -11.9086 35.6604 6.5797C42.1115 18.634 22.3423 83.9305 48.393 82.7136C70.9485 81.6599 44.8471 54.4608 105.478 54.4608C108.496 54.4608 135.023 54.4752 134.763 55.2043',
-				translateValues: [260, 370]
-			},
-			{
-				path: 'M0 0C18.7922 5.26717 42.253 3.56877 61.7533 3.56877C68.4046 3.56877 75.1037 3.74986 81.701 2.82528C83.4944 2.57396 90.7381 1.32984 89.765 1.78439',
-				translateValues: [360, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['north'],
-		lines: [
-			{
-				path: 'M0 25.1758C5.76 11.3284 6.65862 12.2774 10.4488 8.49787C20.2557 -1.28119 40.9186 -5.3721 49.5158 11.4955C69.7675 51.2293 51.4964 181.968 51.4964 231.966',
-				translateValues: [300, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['not guilty'],
-		lines: [
-			{
-				path: 'M0 13.937C7.04917 7.68306 8.82596 5.24486 14.2592 2.47602C22.2666 -1.60461 34.9897 -0.94452 40.471 6.73716C50.5357 20.8421 45.3811 49.6251 42.3582 65.5114C41.4857 70.0969 40.0516 79.3234 40.0516 79.3234C40.0516 79.3234 40.9799 73.4014 42.3582 70.8011C52.7495 51.1976 83.404 30.7263 97.0884 58.0177C118.7 101.119 51.3748 183.061 40.8904 223.467C40.3348 225.608 43.7185 220.065 45.0843 218.325C54.9277 205.784 67.0252 193.693 79.8935 184.676',
-				translateValues: [300, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['nothing'],
-		lines: [
-			{
-				path: 'M0 20.5544C0 32.4368 17.8872 -19.5721 41.8055 8.36105C59.1828 28.6553 22.7549 85.573 47.5352 99.2161C59.2253 105.652 72.7998 91.7178 77.8813 82.5618C80.2627 78.2708 79.3181 63.2184 77.669 67.8406',
-				translateValues: [260, 360]
-			},
-			{
-				path: 'M0.447212 34.0521C-1.50356 32.0017 3.44668 29.0358 5.32805 27.5093C13.7807 20.6512 26.2608 4.59566 36.0986 0',
-				translateValues: [360, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['o'],
-		lines: [
-			{
-				path: 'M0.429231 18.0749C0.567341 17.8677 0.128179 25.5667 0.0405675 27.1437C-0.317389 33.5869 1.59082 45.1513 9.88672 46.1883C25.0307 48.0813 45.7798 33.3513 48.2349 18.0749C52.8101 -10.3931 -2.02838 -2.6224 1.20656 23.2571',
-				translateValues: [330, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ["o'clock"],
-		lines: [
-			{
-				path: 'M0 4.24421C7.71354 2.31582 12.7713 23.5728 26.1928 22.1912C41.3192 20.634 54.566 0 68.1499 0C70.757 0 63.6467 2.62988 61.3592 3.88042C55.6854 6.98208 29.2912 20.7194 35.8939 32.0135C41.379 41.3958 64.7272 37.2414 73.0004 33.1048',
-				translateValues: [320, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['obvious'],
-		lines: [
-			{
-				path: 'M0.00794488 5.25082C0.00794488 6.20282 -0.0933359 5.15953 0.427333 6.13244C4.38595 13.5294 31.8361 59.4614 37.9626 64.6129C39.0526 65.5294 40.1898 62.8374 41.3177 61.968C53.1927 52.8149 67.1918 40.5905 76.3366 28.7605C80.3164 23.6122 91.5603 12.2992 87.2407 4.51614C77.2314 -13.519 38.7464 28.9203 79.0627 19.5036',
-				translateValues: [290, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['obviously'],
-		lines: [
-			{
-				path: 'M0 47.1871C0 48.9778 17.6787 61.5198 19.8063 63.6431C41.7929 85.5855 46.6937 111.175 67.9074 84.6593C76.4335 74.0022 92.7697 63.3993 98.1827 50.7559C110.918 21.0109 54.0154 53.9793 79.5082 59.083C103.295 63.8452 141.538 14.2073 155.055 0',
-				translateValues: [270, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['of', 'officer'],
-		lines: [
-			{
-				path: 'M0 2.49791C17.1808 20.5561 40.9748 18.2048 58.295 0',
-				translateValues: [330, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['of the'],
-		lines: [
-			{
-				path: 'M0 7.87718C0 1.68247 10.5463 29.2784 26.5971 26.9106C41.3095 24.7403 43.8907 11.0079 51.7794 1.33442C52.1688 0.856932 52.9706 -0.429347 53.1941 0.144832C54.1516 2.60493 38.4197 190.002 35.0855 218.038',
-				translateValues: [310, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['on the other hand'],
-		lines: [
-			{
-				path: 'M0 1.30288C6.41135 1.30288 7.59301 20.0883 18.7854 20.0883C33.9897 20.0883 31.5634 6.31472 40.5506 1.8211C47.8443 -1.82576 47.7047 19.0048 55.5789 21.1247C76.9646 26.8824 77.8655 6.15961 89.9109 0.136892C92.178 -0.99666 90.6736 5.215 91.7247 7.52151C97.0314 19.1668 107.02 29.0568 113.49 40.2988C128.845 66.981 145.721 92.9477 164.146 117.513C169.937 125.235 220.502 169.057 220.502 176.331',
-				translateValues: [250, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['only'],
-		lines: [
-			{
-				path: 'M0 1.96882C1.20106 1.68829 7.91034 10.484 11.2472 11.7829C27.8037 18.2282 39.2482 10.9764 49.6573 0.0357336C50.2093 -0.544513 50.3089 5.92457 49.8695 20.7049C49.669 27.4475 47.7323 60.8378 52.8404 66.2068C54.4388 67.8867 55.1656 62.1908 56.448 60.2588C69.4986 40.5976 86.0111 22.5515 101.437 4.7941C101.935 4.22087 103.852 1.39672 104.195 2.11752',
-				translateValues: [280, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['opportunity'],
-		lines: [
-			{
-				path: 'M0 0C0 24.7886 35.0839 17.8478 41.8762 1.18959C42.9091 -1.34368 41.1788 6.6174 40.7444 9.31847C38.0765 25.9101 38.8853 44.6982 40.1785 61.4622C43.4094 103.345 38.1979 145.702 38.1979 187.757C38.1979 199.72 42.7251 210.588 42.7251 222.255C42.7251 226.239 43.2909 230.159 43.2909 234.151C43.2909 236.334 43.008 242.877 43.008 240.694C43.008 236.236 103.97 127.011 111.481 116.183C138.562 77.1476 177.972 46.1996 198.912 2.18092',
-				translateValues: [250, 330]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['or'],
-		lines: [
-			{
-				path: 'M43.2208,0.193563C43.8038,-0.243658 41.7537,0.177213 41.0381,0.314826C20.3549,4.29237 1.50358,19.2725 0.0511315,41.0592C-1.53399,64.8361 34.1302,72.9975 52.1943,67.9797C94.2025,56.3107 83.6067,3.81037 42.2507,2.13377C40.4648,2.06137 32.5549,0.670816 31.337,2.49756',
-				translateValues: [330, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['organisation'],
-		lines: [
-			{
-				path: 'M12.5728,18.2286C-9.23133,28.4142 -1.65845,94.649 27.4275,100.31C91.8007,112.84 137.051,15.8052 59.4713,0.979497C47.2986,-1.34675 31.853,0.217753 21.91,8.41444C18.967,10.8406 9.83131,25.2255 9.1774,24.92',
-				translateValues: [270, 360]
-			},
-			{
-				path: 'M2.82367 27.3586C-7.84825 39.5103 15.155 18.2821 15.3441 18.1392C33.3208 4.55196 67.6736 -12.8191 82.4026 14.2731C96.0269 39.3333 70.748 96.7662 59.6961 119.403C57.8412 123.202 41.8353 159.007 41.0216 156.727',
-				translateValues: [330, 370]
-			},
-			{
-				path: 'M0 13.0464C5.38736 15.563 19.1997 -12.4569 28.224 6.801C34.3041 19.7759 32.256 40.2088 32.256 54.6821',
-				translateValues: [420, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['ought'],
-		lines: [
-			{
-				path: 'M0 0.4768C15.2323 43.1708 41.9192 1.52523 50.6476 3.05425C54.0021 3.64189 37.0799 33.8901 52.3453 40.1299C77.0831 50.2414 78.3503 15.2428 79.5082 2.26119C79.6576 0.586111 152.944 -1.43725 147.416 1.46813',
-				translateValues: [270, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['our', 'hour', 'were', 'world'],
-		lines: [
-			{
-				path: 'M0.395813 4.61741C-1.51304 4.17156 4.00408 6.15589 5.84786 6.82145C12.751 9.31329 19.9108 13.6862 26.6076 16.96C43.1404 25.0424 61.7275 32.3463 79.8699 35.9147C136.71 47.0946 195.594 52.1558 253.077 42.086C277.887 37.7399 298.54 28.5462 321.228 18.4294C329.269 14.844 336.502 9.93047 343.665 4.91129C345.181 3.84924 351.168 -0.574533 349.956 0.0624075',
-				translateValues: [190, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['over and done with'],
-		lines: [
-			{
-				path: 'M0.326062 0.789407C-1.16867 2.36048 2.88093 5.33422 4.28733 6.93563C18.2328 22.8148 43.806 32.0928 59.745 13.4784C61.967 10.8834 72.3508 -3.00612 67.9505 0.591142',
-				translateValues: [220, 380]
-			},
-			{
-				path: 'M0.665479 8.13961C-2.64152 8.13961 7.28256 7.9162 10.5686 7.54481C18.0485 6.69945 25.1319 6.35522 32.6385 6.35522C50.3682 6.35522 68.0507 6.5069 85.5497 4.57083C101.653 2.78918 117.471 35.6591 139.876 37.0863C193.219 40.4844 212.866 2.73868 228.438 0.0107306C229.648 -0.201235 228.71 2.79046 228.438 3.38124C223.514 14.0779 217.723 24.5115 215.14 36.095C214.418 39.3289 215.134 45.81 212.876 45.81',
-				translateValues: [230, 450]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['p', 'page', 'pence ', 'police'],
-		lines: [
-			{
-				path: 'M2.91741 0.0561427C1.2198 -0.509728 1.41686 3.34696 1.10365 5.10877C-1.72955 21.0455 1.73927 40.1208 2.14009 56.1533C3.17248 97.449 9.61097 136.235 7.97005 177.805C7.12835 199.128 7.38811 219.553 8.0996 240.898C8.11979 241.504 7.97005 243.318 7.97005 242.712',
-				translateValues: [320, 340]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['particular'],
-		lines: [
-			{
-				path: 'M0 0C5.90971 6.21153 2.54653 200.702 2.54653 226.915C2.54653 230.284 6.3373 221.343 8.27621 218.587C11.5707 213.905 14.6766 209.093 17.8257 204.312C27.1765 190.116 32.7735 174.251 43.2909 160.595C53.4663 147.383 153.503 4.74085 156.611 4.01487',
-				translateValues: [280, 340]
-			},
-			{
-				path: 'M0 1.55418C40.1736 1.55418 80.227 0.661986 120.323 0.661986C121.543 0.661986 137.725 0.724773 137.725 0.21589C137.725 -0.575141 130.481 1.02281 127.751 1.55418C107.516 5.49264 82.0919 16.4392 83.1865 40.2159C84.3579 65.661 130.017 42.958 139.21 40.8107',
-				translateValues: [380, 320]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['people'],
-		lines: [
-			{
-				path: 'M110.041 0.137257C115.486 -1.22399 101.906 7.90711 97.4297 11.2935C85.4847 20.3308 75.3193 31.0244 64.9312 41.7305C28.6983 79.0726 13.7405 101.5 1.87432 152.201C-13.4364 217.62 68.3917 291.668 125.805 239.996',
-				translateValues: [250, 300]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['perfect'],
-		lines: [
-			{
-				path: 'M8.81846 0C7.11425 3.5825 -2.94839 282.354 0.850084 291.227C2.08152 294.103 4.54469 286.175 6.30213 283.586C11.5848 275.805 16.6496 267.933 22.2389 260.37C37.3774 239.886 55.1077 221.549 70.4686 201.155C99.3223 162.847 185.454 76.8483 177.203 26.7423C170.083 -16.4943 125.943 5.35774 109.052 32.7667C82.8529 75.2805 85.3568 135.978 85.3568 182.935',
-				translateValues: [250, 320]
-			}
-		]
-	},
-	{
-		letterGroupings: ['prm'],
-		specialOutlineMeanings: ['permanent'],
-		lines: [
-			{
-				path: 'M6.84878 0.280825C6.84878 -0.816001 6.57412 1.61437 6.63909 2.19099C8.96175 22.8063 12.094 46.5413 11.4995 67.7857C10.4014 107.028 7.30479 148.314 2.90204 187.391C1.49084 199.917 4.93912 209.612 2.02582 221.86C0.562991 228.01 -1.40262 228.486 1.47911 237.194C16.7141 242.317 36.6294 190.822 45.0131 179.542C66.6115 150.485 90.1818 123.588 107.921 91.8218C140.842 32.8696 149.711 -7.81471 230.592 4.98277C254.487 8.76357 275.64 22.5387 285.952 44.2146',
-				translateValues: [220, 350]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['prejudice'],
-		lines: [
-			{
-				path: 'M12.1669 0C12.1669 65.5296 10.0801 134.808 1.13191 200.644C-1.12177 221.066 0.325423 238.894 2.54665 255.366',
-				translateValues: [270, 320]
-			},
-			{
-				path: 'M102.993 0.223448C109.192 -2.6724 104.408 23.4539 104.408 27.9806C104.408 47.5876 105.121 67.138 106.388 86.6671C110.267 146.462 74.9921 262.132 0 262.132',
-				translateValues: [320, 310]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['preliminary'],
-		lines: [
-			{
-				path: 'M9.58989 0C9.58989 18.4387 7.47151 45.6889 6.22484 64.0336C1.98054 126.488 1.90989 175.155 1.90989 237.12C1.90989 252.575 0 283.014 0 298.185',
-				translateValues: [250, 310]
-			},
-			{
-				path: 'M94.1507 0C83.9213 6.12588 63.9292 18.3062 49.9907 32.64C35.8225 47.2102 27.7528 63.943 23.1107 71.04C0.228076 106.024 0 160.048 0 201.65C0 202.659 10.1428 188.083 10.6307 187.2C29.4894 166.369 42.2986 158.073 70.1507 156.48C97.1443 154.936 125.061 175.209 135.532 199.667C135.61 199.849 135.45 199.253 135.532 199.072C140.278 188.54 152.064 172.401 162.412 166.358C163.354 165.808 163.367 165.552 164.109 164.772',
-				translateValues: [290, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['prepare'],
-		lines: [
-			{
-				path: 'M18.1086 0C6.0156 21.1844 0 234.952 0 277.77',
-				translateValues: [320, 320]
-			},
-			{
-				path: 'M14.9341 0C-4.16687 80.3061 0.503814 194.816 0.503814 278.166',
-				translateValues: [360, 320]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['principle', 'principal'],
-		lines: [
-			{
-				path: 'M7.73804 0C7.73804 14.6137 7.64666 29.236 6.60625 43.8166C3.20192 91.5258 2.30112 132.613 6.36372 181.007C7.07778 189.513 7.56004 202.373 7.39125 215.581C7.19452 230.976 6.11331 246.845 3.48372 256.847C3.23876 257.779 0.179657 289.949 0.0984567 289.864C-2.27372 287.37 38.9884 231.456 40.9237 229.007C62.164 202.135 66.1721 193.993 83.285 164.56C110.258 118.169 122.89 86.9487 158.549 45.0062C166.345 35.8364 171.095 21.8917 180.053 14.4734C182.051 12.8185 189.407 1.12738 192.22 1.78439C200.094 3.6235 177.224 59.3907 177.224 70.7807C177.224 74.1495 180.972 64.9816 183.731 63.0484C190.282 58.4578 199.709 51.6418 207.499 57.1004C226.715 70.5655 211.743 121.731 211.743 141.165',
-				translateValues: [260, 320]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['probable'],
-		lines: [
-			{
-				path: 'M3.23779 1.8224C3.23779 0.600826 3.25933 2.22398 3.23779 2.41719C2.07585 12.8388 0.691264 23.0099 0.691264 33.4953C0.691264 49.6475 -0.230171 65.6576 0.0546329 81.8224C0.964125 133.443 1.75232 183.746 1.75232 234.685C1.75232 237.272 9.48429 211.731 12.1506 203.012C17.0422 187.017 23.3869 171.379 31.0374 156.469C49.8151 119.874 83.0029 22.0637 114.436 0.0380105C115.331 -0.589087 112.513 6.49505 102.977 24.8707C96.2026 37.9243 89.6388 50.9794 83.0289 64.1273C63.1939 103.582 52.8991 164.344 84.9388 201.228C114.678 235.463 153.286 186.38 141.175 153.495C133.203 131.851 104.788 129.06 85.5755 134.313C79.596 135.948 75.1049 138.075 69.8719 141.302C68.3907 142.215 62.4702 146.696 62.8689 145.019',
-				translateValues: [290, 330]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['probably'],
-		lines: [
-			{
-				path: 'M17.5427 0C17.5427 25.7643 11.2762 52.536 9.90316 78.513C7.18264 129.983 4.64998 182.235 1.69768 233.953C0.527288 254.456 4.15092 271.486 0 291.846',
-				translateValues: [250, 290]
-			},
-			{
-				path: 'M93.468 0C84.3356 8.8181 69.629 19.9231 68.508 21.12C52.1229 38.6153 53.7049 33.9697 40.9005 55.4035C18.1066 93.5589 -5.02934 164.633 2.70257 209.059C14.6419 277.661 120.987 319.24 141.63 232.454C144.434 220.666 144.646 204.827 137.103 194.586C109.86 157.6 53.1174 163.922 32.695 201.128C19.4735 225.216 8.69935 250.174 2.98552 276.866C2.39095 279.643 -1.23865 291.283 0.438989 286.581',
-				translateValues: [290, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['problem'],
-		lines: [
-			{
-				path: 'M3.02306 0.0255769C4.5445 -0.329788 2.51202 3.10981 2.17421 4.63524C0.677497 11.394 1.32537 18.5922 1.32537 25.4531C1.32537 46.4352 -0.710645 68.2595 0.264319 89.0962C1.57406 117.088 7.93421 142.925 9.81379 170.583C11.6344 197.374 5.994 230.783 5.994 257.126',
-				translateValues: [290, 340]
-			},
-			{
-				path: 'M69.5288 1.62588C69.5288 -5.85878 62.0004 14.5747 57.8573 20.808C47.568 36.2879 37.9149 52.192 31.3309 69.5813C17.5144 106.073 -19.0469 203.346 12.4442 238.652C61.7291 293.907 138.629 231.779 114.305 177.239C101.101 147.631 58.9082 151.386 33.4531 158.206C25.3789 160.369 13.6088 166.064 9.47327 173.671C7.39663 177.49 2.47032 189.022 2.47032 184.674',
-				translateValues: [330, 340]
-			}
-		]
-	},
-	{
-		letterGroupings: ['pf'],
-		specialOutlineMeanings: ['profit'],
-		lines: [
-			{
-				path: 'M1.2378 0.297159C0.509785 -1.23324 0.20661 3.56204 0.106013 5.25379C-0.53707 16.0686 1.88973 23.8586 3.50138 34.3988C6.48112 53.8863 4.69798 75.6868 5.48201 95.4645C6.71847 126.655 -8.63763 279.32 23.0247 291.152C71.328 309.204 49.3546 155.915 25.5713 127.583C23.3238 124.906 8.42386 103.357 6.89675 106.567',
-				translateValues: [290, 300]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['proportion'],
-		lines: [
-			{
-				path: 'M0.959123 1.28152C0.959123 -5.07715 1.86999 13.9666 2.23239 20.315C2.97472 33.319 2.42373 45.9691 2.02018 58.9767C1.01567 91.3546 14.3137 129.661 5.83996 160.092C2.22896 173.06 2.05567 186.621 0.534702 199.943C0.296906 202.026 -0.539304 219.822 0.534702 219.572',
-				translateValues: [280, 370]
-			},
-			{
-				path: 'M4.42802 0.328712C4.42802 -0.84005 4.04029 1.45576 4.0036 1.8157C3.32209 8.50138 2.51813 15.4909 2.51813 22.1875C2.51813 33.6025 1.40107 44.7321 1.03266 56.0908C-0.249032 95.6071 6.49527 131.223 3.15476 169.845C1.87908 184.595 5.80607 202.911 1.45708 217.132C1.09869 218.304 1.09891 224.762 0.183814 224.121C-1.43559 222.986 8.15205 206.494 9.09666 204.641C30.8416 161.978 73.8981 117.736 85.7046 71.4068C90.451 52.7823 97.6538 38.4069 105.865 21.1466C107.936 16.7914 117.922 0.221065 116.475 2.85659',
-				translateValues: [310, 370]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['prosecution'],
-		lines: [
-			{
-				path: 'M27.5574 0C28.0508 0.518576 28.053 0.763608 27.9819 1.48699C22.7416 54.7303 23.9499 105.69 23.9499 159.108C23.9499 166.533 20.649 205.271 26.2842 209.219C36.124 216.114 61.1778 194.466 48.3541 188.996C-1.1632 167.876 -33.4453 275.783 58.1157 233.011',
-				translateValues: [300, 320]
-			},
-			{
-				path: 'M1.6445 11.0438C-6.31909 11.0438 16.8874 -3.80338 23.29 0.932235C34.9076 9.52511 32.0623 45.0938 26.4731 56.843',
-				translateValues: [380, 350]
-			}
-		]
-	},
-	{
-		letterGroupings: ['pb'],
-		specialOutlineMeanings: ['public'],
-		lines: [
-			{
-				path: 'M24.5049 0C29.1302 4.86162 16.2075 71.2695 14.6017 83.2714C12.4447 99.3935 17.1081 115.926 15.1676 132.243C8.8984 184.958 -3.7021 230.176 4.13264 282.33C10.4646 324.48 116.975 289.852 87.6021 210.954C78.9998 187.848 48.9683 175.581 28.7491 190.929C21.0001 196.812 16.3063 203.777 11.2063 211.946C9.17207 215.204 -1.47027 235.718 0.171377 235.143',
-				translateValues: [270, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['q', 'question', 'equal'],
-		lines: [
-			{
-				path: 'M28.0655 2.07287C30.3589 2.07287 23.5694 14.1622 23.0129 15.417C17.305 28.2856 14.4203 42.4492 10.5756 55.9676C2.86648 83.0739 -14.2874 162.058 23.9197 175.158C58.2442 186.926 86.5571 44.2607 87.272 18.5263C87.3057 17.3106 88.5096 0.536126 86.3651 0',
-				translateValues: [280, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['r', 'are'],
-		lines: [
-			{
-				path: 'M0.0175922 182.543C-0.689805 183.604 20.0661 164.256 38.884 140.826C64.3145 109.162 88.9611 77.9544 115.839 47.5466C128.695 33.0027 142.295 19.2599 155.095 4.66397C156.393 3.18315 160.692 0 158.722 0',
-				translateValues: [270, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['recognise'],
-		lines: [
-			{
-				path: 'M0.775046 208.198C-3.49505 220.167 10.3742 202.862 30.2016 187.182C47.5392 173.472 62.7536 156.74 81.415 144.754C109.602 126.649 231.863 2.24331 250.901 0.019972C255.234 -0.486108 250.862 8.75446 251.183 13.1055C252.107 25.6166 253.031 40.3431 256.842 52.362C257.362 53.9997 257.105 57.1204 258.823 57.1204C261.339 57.1204 257.537 51.7515 255.428 50.3793C248.837 46.0908 242.438 43.2212 235.621 48.595C224.911 57.0381 218.809 98.534 226.284 110.057C227.605 112.094 229.29 113.562 231.377 114.816C233.436 116.052 238.679 117.234 241.28 116.996C280.715 113.392 227.982 72.0673 227.982 110.652',
-				translateValues: [170, 250]
-			}
-		]
-	},
-	{
-		letterGroupings: ['rc'],
-		specialOutlineMeanings: ['recommend', 'recent'],
-		lines: [
-			{
-				path: 'M0 175.947C0 173.419 48.2349 146.657 64.3661 136.232C130.011 93.8118 191.573 44.7876 255.487 0.00190352C255.696 -0.144105 242.183 8.14542 235.497 12.4704C223.707 20.0961 216.203 33.2198 217.483 47.5671C219.03 64.9181 259.314 51.2226 263.176 43.103',
-				translateValues: [210, 290]
-			}
-		]
-	},
-	{
-		letterGroupings: ['rf'],
-		specialOutlineMeanings: ['reference', 'referee'],
-		lines: [
-			{
-				path: 'M0.237329 228.204C-1.92574 230.478 11.3897 217.762 12.1211 216.903C25.7769 200.869 39.2225 184.48 53.4314 168.923C75.1813 145.11 194.364 29.9686 175.948 8.12983C145.326 -28.1825 69.2235 67.4726 64.1834 92.1943C60.6805 109.376 57.9586 127.975 57.9586 145.528C57.9586 148.725 57.6029 164.986 58.2415 156.036',
-				translateValues: [270, 230]
-			}
-		]
-	},
-	{
-		letterGroupings: ['rg'],
-		specialOutlineMeanings: ['regular'],
-		lines: [
-			{
-				path: 'M0 112.929C1.70764 114.724 5.16474 109.822 8.27621 107.279C22.613 95.5583 38.4062 84.6393 52.2038 72.3344C60.2178 65.1873 116.395 27.807 117.14 25.1968C117.8 22.8863 95.7036 18.7076 101.861 11.5165C112.731 -1.17808 141.364 -3.77642 154.489 5.71729C191.89 32.7701 139.515 94.7191 120.323 115.606C113.353 123.192 105.938 130.49 97.829 136.87C95.9064 138.382 80.0771 150.631 83.6109 146.386',
-				translateValues: [250, 340]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['remark'],
-		lines: [
-			{
-				path: 'M0 192.96C8.78743 183.223 23.4134 165.313 32.9836 156.372C66.1451 125.39 97.484 90.4293 126.922 56.0499C137.363 43.8563 148.801 31.4668 160.593 20.5604C161.701 19.5357 168.252 9.93165 180.48 0',
-				translateValues: [190, 260]
-			},
-			{
-				path: 'M0 58.9906C19.2727 50.5684 45.209 35.5824 64.4716 27.2858C120.896 2.98289 175.254 -3.03887 237.11 1.29544C272.151 3.75081 300.363 16.1576 330.483 32.8196C340.617 38.4257 349.193 45.3587 358.777 51.6548',
-				translateValues: [200, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['remember'],
-		lines: [
-			{
-				path: 'M1.70552 197.756C-4.63901 194.792 7.92601 192.344 19.5312 183.283C41.1096 166.435 59.4874 146.465 78.9502 127.372C112.59 94.3717 146.118 62.4223 178.265 28.0412C181.115 24.9923 196.413 0.284046 201.466 0.284046C202.004 0.284046 200.635 48.4625 203.447 48.4625C211.158 48.4625 213.293 36.4545 219.292 31.61C234.185 19.5827 256.894 12.5557 275.882 9.00772C291.698 6.05224 352.277 12.766 352.277 40.7302C352.277 46.51 359.716 31.8532 363.878 27.8429C371.361 20.6329 380.785 14.7504 389.909 9.80078C421.576 -7.37792 457.416 0.574665 488.658 14.1626C497.514 18.0143 506.594 22.7151 514.406 28.4377C519.056 31.8444 532.709 42.4521 527.988 39.144',
-				translateValues: [100, 260]
-			}
-		]
-	},
-	{
-		letterGroupings: ['rs'],
-		specialOutlineMeanings: ['residents'],
-		lines: [
-			{
-				path: 'M0 196.286C0 200.299 5.80905 190.733 8.93927 188.221C19.38 179.845 28.6715 170.18 39.0607 161.695C65.1603 140.38 95.2578 125.437 120.874 103.104C130.533 94.6835 207.778 20.4498 204.437 8.75585C197.024 -17.1922 155.672 22.5618 187.142 24.9826C191.086 25.2859 193.882 24.4459 197.053 22.0676',
-				translateValues: [260, 260]
-			}
-		]
-	},
-	{
-		letterGroupings: ['rsp'],
-		specialOutlineMeanings: ['respect'],
-		lines: [
-			{
-				path: 'M0 110.494C0.217681 110.952 23.0431 94.2276 46.2619 77.0369C55.2199 70.4047 158.985 15.6721 127.539 1.34919C93.763 -14.0343 101.411 106.856 100.376 124.026C99.6037 136.826 98.2535 149.896 98.2535 162.688C98.2535 164.488 98.6779 172.057 98.6779 170.569',
-				translateValues: [260, 350]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['s', 'south', 'southern', 'survive'],
-		lines: [
-			{
-				path: 'M0.281887 16.9918C6.59681 30.2667 35.1664 38.5347 33.833 16.1102C31.6769 -20.148 -11.6081 15.0411 3.00791 22.7223',
-				translateValues: [350, 423]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['satisfactory'],
-		lines: [
-			{
-				path: 'M10.2259 27.6411C36.9738 27.6411 45.3197 -15.2597 13.3383 5.83195C8.39312 9.09328 1.7341 12.8523 0.32274 19.1157C-0.317336 21.9563 -0.171534 25.4919 2.30337 27.4429C20.392 41.7022 99.0849 16.2866 109.54 26.055C111.456 27.8451 105.772 29.7433 103.599 31.2099C91.8039 39.1684 61.7698 59.4109 59.4587 76.4144C56.4341 98.6679 108.59 65.4432 78.4162 63.3289C74.1754 63.0317 70.8625 64.915 67.0983 64.915',
-				translateValues: [280, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['school'],
-		lines: [
-			{
-				path: 'M3.13243 21.3862C8.92872 21.8692 9.19664 27.8589 15.7438 29.6321C32.0016 34.0352 61.5247 7.94753 39.9964 0.771442C27.0132 -3.5563 -15.9669 11.1165 6.28527 25.0241',
-				translateValues: [320, 270]
-			}
-		]
-	},
-	{
-		letterGroupings: ['sv'],
-		specialOutlineMeanings: ['several'],
-		lines: [
-			{
-				path: 'M7.04257 25.8238C12.3251 20.2715 21.6145 20.3924 25.2927 12.2922C33.2069 -5.13684 1.82337 -3.39183 0.0396228 13.4818C-0.808361 21.5034 12.1818 29.3527 17.2287 33.8535C32.6548 47.6109 47.2891 64.5913 65.1882 75.0432C70.2143 77.978 72.0496 65.5681 76.2232 61.5115C93.6897 44.5348 108.005 24.5468 123.334 5.74945',
-				translateValues: [290, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['sh', 'shall'],
-		lines: [
-			{
-				path: 'M39.6437 0.976551C42.8896 -0.105423 30.4322 3.54623e-05 30.1215 0.00489157C20.0612 0.162083 8.79953 4.22885 6.41296 14.9685C2.76222 31.3968 33.4575 20.1484 39.4494 28.2802C50.3535 43.0786 29.6163 49.2958 19.4332 51.1142C12.3874 52.3724 6.50532 50.5312 0 50.5312',
-				translateValues: [360, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['signature'],
-		lines: [
-			{
-				path: 'M36.935 8.73667C20.9664 42.305 90.2917 5.29551 62.4003 0.409529C25.0412 -6.13499 14.5196 67.5916 0.434835 90.6202C-1.7969 94.2691 5.14274 83.425 8.07441 80.3104C12.9911 75.0871 18.9028 69.5506 24.7683 65.4405C40.8825 54.1491 80.9199 41.966 93.8075 65.044C115.398 103.706 71.1739 162.587 49.6677 192.727C42.5902 202.645 35.4991 213.145 27.3148 222.268C24.172 225.772 14.049 238.9 14.5822 238.526',
-				translateValues: [300, 370]
-			}
-		]
-	},
-	{
-		letterGroupings: ['sg'],
-		specialOutlineMeanings: ['significant'],
-		lines: [
-			{
-				path: 'M25.8897 9.65375C60.6807 9.65375 36.037 61.2999 13.1571 36.717C-2.17566 20.2431 82.8596 -33.4837 96.1314 30.769C108.311 89.7328 36.1943 147.976 0 183.483',
-				translateValues: [300, 360]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['social'],
-		lines: [
-			{
-				path: 'M2.82947 14.3783C-6.04731 33.0386 51.0945 50.1866 38.4808 10.413C32.5371 -8.32895 0 0.584262 0 18.5419',
-				translateValues: [290, 420]
-			},
-			{
-				path: 'M40.4615 1.48114C40.4615 -4.26619 5.39966 7.7918 4.24421 19.1267C3.53161 26.1174 45.2777 20.0291 38.1979 46.4873C33.4463 64.2449 0.68453 53.4906 0 55.4093',
-				translateValues: [350, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['society'],
-		lines: [
-			{
-				path: 'M13.572 7.50757C15.4622 8.39055 13.8497 13.8115 14.1379 16.033C16.7076 35.8395 59.4797 26.8855 56.2971 11.2746C51.7684 -10.9385 -4.37754 -0.647137 0.27351 44.9797C3.38642 75.5173 48.688 73.0977 64.7855 54.2982',
-				translateValues: [320, 390]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['something'],
-		lines: [
-			{
-				path: 'M15.9741 11.1927C58.9806 11.1927 28.0766 63.7165 4.11143 45.0579C-9.59679 34.3851 14.6937 15.0422 22.7842 10.8848C53.0539 -4.66948 93.9977 -2.78505 124.496 11.6545C127.718 13.1799 164.893 44.5503 159.425 45.8275',
-				translateValues: [240, 410]
-			},
-			{
-				path: 'M0 36.8105C15.2877 22.5275 25.5309 -0.791118 24.6041 0.0206417',
-				translateValues: [420, 420]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['station'],
-		lines: [
-			{
-				path: 'M0.17357 15.7198C-2.39491 15.7198 24.1628 -10.4202 38.7959 4.71608C52.9158 19.3215 45.5866 64.266 45.5866 83.2291',
-				translateValues: [330, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['substantial'],
-		lines: [
-			{
-				path: 'M56.9197 12.0685C56.9197 40.2162 115.223 3.74307 82.102 0.172589C36.5851 -4.73423 1.43722 96.3437 0.0472451 130.433C-1.95217 179.468 60.0024 273.119 92.2881 186.74C109.663 140.254 56.6134 106.708 17.8729 130.235C-23.6715 155.464 49.5668 180.688 43.9041 138.363C42.4122 127.213 25.741 128.525 17.307 129.838C16.912 129.9 15.9431 129.767 16.1752 129.442C17.445 127.662 20.0387 127.343 22.1171 126.666C60.8767 114.034 105.351 119.481 145.482 116.356C146.331 116.29 147.442 115.541 148.029 116.158C150.298 118.543 146.96 139.322 146.614 143.32C146.237 147.675 147.643 160.353 145.765 156.406',
-				translateValues: [270, 230]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['success'],
-		lines: [
-			{
-				path: 'M15.405 5.64996C15.405 29.5916 57.0753 7.23805 38.6067 1.48639C8.86737 -7.77528 -3.77854 28.4123 0.974663 50.0614C5.10913 68.8924 31.0062 64.7056 42.285 56.8024C44.2332 55.4372 45.9633 52.096 45.9633 56.2076',
-				translateValues: [290, 390]
-			},
-			{
-				path: 'M3.67832 9.46837C3.08701 8.22536 2.9071 10.0596 2.82947 10.658C0.158405 31.2462 24.7965 28.5408 33.3878 17.0025C34.8456 15.0445 37.0155 12.2716 36.7832 9.66664C35.512 -4.58502 4.08517 -1.78352 0 9.66664',
-				translateValues: [360, 430]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['successful'],
-		lines: [
-			{
-				path: 'M12.9285 129.764C12.9285 152.89 65.7301 117.885 31.821 117.142C27.6669 117.051 23.8737 117.617 19.9583 118.989C-16.0609 131.609 1.03559 191.682 36.6539 190.722C89.9294 189.286 223.566 32.2673 186.695 3.84756C141.47 -31.0112 48.8542 182.336 135.29 187.181C150.587 188.039 166.329 181.396 175.272 168.863',
-				translateValues: [270, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['such'],
-		lines: [
-			{
-				path: 'M19.4326 7.08616C11.7425 28.6405 49.5808 24.8295 48.1607 1.94341C47.9885 -0.832424 34.4052 0.180181 33.9015 0.180181C13.9265 0.180181 3.31525 21.4668 0.350436 38.3835C-4.67762 67.0726 46.0638 62.4112 46.0638 55.428',
-				translateValues: [320, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['super'],
-		lines: [
-			{
-				path: 'M0 22.1806C5.00418 24.1822 12.2953 21.5898 16.4918 19.149C43.6019 3.38087 10.4328 -9.66259 3.39537 9.69046C0.465899 17.7465 0.242526 27.6571 0.242526 36.1258C0.242526 48.4563 0.727579 60.9688 0.727579 73.1111',
-				translateValues: [360, 300]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['supermarket'],
-		lines: [
-			{
-				path: 'M3.56937 18.1353C3.56937 19.8511 7.38643 19.2049 8.87463 19.0275C39.0919 15.4234 25.268 -10.1193 6.96474 4.45496C-3.14462 12.5047 0.429474 40.3597 1.44727 51.2951C1.91287 56.2977 4.87879 70.8507 2.72053 66.3137',
-				translateValues: [350, 280]
-			},
-			{
-				path: 'M0 73.8915C86.0142 13.6201 207.436 -17.524 311.737 10.2483C345.237 19.1684 381.457 31.8589 412.325 47.7205C418.486 50.8866 463.184 73.961 452.008 75.5272',
-				translateValues: [140, 380]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['t', 'to'],
-		lines: [
-			{
-				path: 'M3.11026 6.11941C-6.51375 6.11941 9.13197 4.82471 10.6244 4.6943C35.0405 2.56086 59.2508 1.65983 83.6933 0.0303372C87.2889 -0.209374 107.886 1.02922 103.774 1.71455',
-				translateValues: [290, 310]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['technical'],
-		lines: [
-			{
-				path: 'M0.605444 4.48674C-2.42502 3.98166 6.74941 4.57007 9.82144 4.608C17.1775 4.69882 24.5349 4.65704 31.8913 4.608C51.8837 4.47472 71.8726 0 91.7953 0C96.9006 0 81.4851 0.980376 76.7587 2.91032C67.1235 6.84469 52.4484 12.5587 50.8084 24.8589C48.4506 42.5423 77.9043 35.5193 87.4299 32.7411',
-				translateValues: [290, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['telephone'],
-		lines: [
-			{
-				path: 'M0.765169 5.74969C-3.09227 4.84871 8.68906 5.85375 12.649 5.74969C26.253 5.39222 39.7869 4.93493 53.3934 4.75837C86.3512 4.3307 119.297 0 152.142 0C154.944 0 147.238 2.8849 145.351 4.95663C132.425 19.147 120.445 35.0485 108.851 50.3594C89.3906 76.0584 19.255 189.61 56.5058 215.713C63.5801 220.67 72.5851 210.3 76.3121 206.196C94.8586 185.774 103.239 150.386 98.9479 123.321C98.2961 119.21 88.8616 83.4639 88.4789 83.8662',
-				translateValues: [280, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['thank you'],
-		lines: [
-			{
-				path: 'M0 0C11.9777 3.64181 31.0719 2.56644 36.5002 2.97398C39.5474 3.20275 97.1379 4.42274 98.7486 5.55143C100.697 6.91673 98.8232 10.311 98.7486 12.689C97.7474 44.6114 98.7486 76.0824 98.7486 107.856C98.7486 123.575 102.993 178.342 98.1827 212.342',
-				translateValues: [260, 240]
-			},
-			{
-				path: 'M0.599191 0C-3.05744 21.413 10.4848 51.5819 28.611 48.3767C39.9453 46.8876 54.0593 35.9695 57.0775 26.1001C59.0574 19.6258 59.4841 6.80202 57.4716 0.793061',
-				translateValues: [330, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['thank you for your'],
-		lines: [
-			{
-				path: 'M1.06074 2.17702C-0.0861627 2.17702 -0.479882 2.52945 0.777797 2.97008C11.8527 6.85025 26.5338 4.5562 38.1268 4.5562C43.7927 4.5562 121.719 -1.10744 123.577 0.194369C124.397 0.769258 123.577 4.90647 123.577 6.53886C123.577 60.2125 126.406 114.547 126.406 167.53C126.406 182.617 123.011 197.241 123.011 212.14C123.011 217.22 119.577 231.454 119.899 230.777',
-				translateValues: [230, 230]
-			},
-			{
-				path: 'M0.282947 0.770786C0.282947 -0.913164 0 5.98098 0 6.91701C0 11.9532 0.248978 16.2909 1.41474 21.1921C5.15084 36.8997 10.2598 46.7683 28.2947 46.7683C38.9834 46.7683 61.4243 30.4726 65.0779 21.1921C66.7287 16.999 68.4733 13.6342 68.4733 9.09793C68.4733 6.25456 68.1903 -2.27085 68.1903 0.57252C68.1903 17.2544 73.2469 33.1722 74.4152 49.544C74.4971 50.6921 74.9969 54.0501 74.4152 48.751',
-				translateValues: [320, 460]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['that'],
-		lines: [
-			{
-				path: 'M1.34023 0.762816C-5.90631 -1.6527 18.5895 2.43992 18.7006 2.44703C39.0792 3.75126 60.1731 -0.958616 80.6277 0.244597C85.659 0.54056 78.8631 10.1942 77.3888 15.0138C72.3951 31.3393 69.8896 46.9932 68.4495 63.9855C64.7601 107.521 62.49 152.099 62.49 195.613',
-				translateValues: [270, 260]
-			},
-			{
-				path: 'M0 5.18219C21.0849 -1.84611 49.1932 0.617036 71.3846 0.259109C77.6709 0.157717 107.43 0.567971 101.182 0',
-				translateValues: [280, 350]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['that the'],
-		lines: [
-			{
-				path: 'M0.242538 0.489996C-1.29693 -0.229153 4.92435 0.0401536 6.75033 0.0934655C19.0418 0.452332 31.2324 0.886527 43.5335 0.886527C53.2206 0.886527 132.95 1.09043 133.511 1.67959C134.657 2.88387 129.969 17.9404 129.832 18.5321C118.051 69.7144 114.869 123.94 112.29 176.351C111.358 195.287 101.821 213.847 101.821 232.46',
-				translateValues: [200, 230]
-			},
-			{
-				path: 'M0 2.78121C44.1741 13.099 101.494 -11.1725 144.586 6.94478C147.378 8.1185 140.057 26.7112 139.21 29.547C132.865 50.7884 135.73 74.7696 131.288 96.5607C124.342 130.629 128.331 216.779 116.857 240.898',
-				translateValues: [250, 300]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['the'],
-		lines: [
-			{
-				path: 'M1.34023 0.762816C-5.90631 -1.6527 18.5895 2.43992 18.7006 2.44703C39.0792 3.75126 60.1731 -0.958616 80.6277 0.244597C85.659 0.54056 78.8631 10.1942 77.3888 15.0138C72.3951 31.3393 69.8896 46.9932 68.4495 63.9855C64.7601 107.521 62.49 152.099 62.49 195.613',
-				translateValues: [270, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['their', 'there'],
-		lines: [
-			{
-				path: 'M0 3.11105C44.7028 -7.33024 185.84 11.9497 244.254 10.8434C271.873 10.3203 299.454 11.2331 327.016 9.8025C337.353 9.26596 347.687 8.4475 357.999 7.86942C359.158 7.80445 372.988 7.31511 366.275 6.53113',
-				translateValues: [170, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['therefore'],
-		lines: [
-			{
-				path: 'M0 13.8837C24.3942 13.8837 92.7292 10.4067 92.736 10.4637C97.0841 47.0252 103.067 93.7831 93.797 129.869C92.1322 136.35 71.4245 185.908 58.1457 171.951C32.6262 145.128 112.711 59.3907 129.024 43.0287C140.797 31.2211 154.636 21.2684 165.1 8.23317C166.485 6.50757 173.34 -2.15078 169.556 0.500822',
-				translateValues: [280, 284]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['they'],
-		lines: [
-			{
-				path: 'M0 0C21.9863 5.13538 87.7999 -3.18891 101.224 3.86617C106.439 6.60647 97.9117 15.1881 96.768 20.9665C94.0661 34.6179 94.721 48.5058 93.5848 62.3049C90.6152 98.3727 94.9692 139.6 88.704 174.721C88.034 178.477 93.9807 169.209 96.5558 166.394C105.243 156.898 112.926 146.673 122.87 138.439C124.776 136.861 131.399 129.375 129.661 135.465',
-				translateValues: [280, 290]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['time'],
-		lines: [
-			{
-				path: 'M0 20.8179C1.85873 22.7715 19.2539 12.6149 21.4333 11.7472C40.4179 4.18878 59.3755 0 79.7912 0C98.5415 0 110.32 10.6433 122.87 23.3457C123.346 23.8274 127.172 26.3197 125.204 26.3197',
-				translateValues: [290, 300]
-			}
-		]
-	},
-	{
-		letterGroupings: ['td'],
-		specialOutlineMeanings: ['to do'],
-		lines: [
-			{
-				path: 'M0 10.3074C25.4523 7.42737 83.1479 1.81281 90.9474 1.33389C95.9698 1.0255 104.064 0.231893 105.984 0',
-				translateValues: [260, 260]
-			},
-			{
-				path: 'M0 16.0067C0 15.3142 7.22242 15.1307 8.97347 14.9154C48.0122 10.1155 87.3644 7.52584 126.114 0.606316C127.015 0.445434 126.955 0.712246 126.599 0',
-				translateValues: [300, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['to the'],
-		lines: [
-			{
-				path: 'M0.152141 1.46936C-2.14516 1.46936 22.2319 1.90745 28.8802 1.61629C46.0647 0.863705 63.3476 1.19944 80.465 0',
-				translateValues: [260, 280]
-			},
-			{
-				path: 'M0.619553 1.1607C-2.45752 1.1607 6.76911 1.47727 9.8461 1.45457C16.1573 1.40802 81.1345 -2.15332 85.1263 2.04232C88.6449 5.74064 83.2186 32.9576 82.4003 39.5109C75.3589 95.8999 66.6732 152.245 66.6732 208.781',
-				translateValues: [320, 250]
-			}
-		]
-	},
-	{
-		letterGroupings: ['tg'],
-		specialOutlineMeanings: ['together'],
-		lines: [
-			{
-				path: 'M0 36.9621C11.7736 32.0121 64.4031 32.8134 80.8522 32.3524C86.8657 32.1838 98.8901 31.7576 98.8901 31.7576C98.8901 31.7576 93.3733 26.0629 92.9482 22.3896C92.2141 16.0451 99.7216 10.1243 103.983 6.92488C117.051 -2.88631 139.071 -1.13546 153.004 5.14049C217.231 34.072 95.2049 152.245 66.4219 167.371',
-				translateValues: [280, 310]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['tomorrow'],
-		lines: [
-			{
-				path: 'M1.08418 50.3673C-5.64252 52.724 21.0998 54.7685 21.8808 54.8282C39.2888 56.1589 85.9846 56.5785 87.6661 53.0438C110.671 4.68329 215.515 -7.0697 260.618 3.67579C271.691 6.31391 282.223 9.76894 292.874 13.7873C305.884 18.6964 319.698 24.0484 330.223 33.2669C332.226 35.0217 348.534 51.7903 345.714 52.449',
-				translateValues: [190, 250]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['u'],
-		lines: [
-			{
-				path: 'M3.31531 1.07448C3.31531 -4.17671 0.410944 11.251 0.0764433 16.4915C-0.519162 25.8226 2.20131 40.9541 14.0683 40.3295C27.428 39.6264 31.0542 23.2412 30.0036 12.7344C29.8592 11.2905 28.0602 -1.07671 28.0602 2.49958',
-				translateValues: [330, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['v', 'very', 'have ', 'village ', 'versus'],
-		lines: [
-			{
-				path: 'M2.20912 5.47028C-8.15356 2.01605 20.952 39.4146 27.3427 49.0006C27.7215 49.5688 28.3431 51.3473 28.5087 50.6849C32.5792 34.4031 43.0715 15.95 52.4763 1.84275C53.6505 0.0815513 54.3849 -1.05021 53.3832 1.45409',
-				translateValues: [320, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['vandalism'],
-		lines: [
-			{
-				path: 'M0 0C8.60514 0 44.5346 39.1643 50.746 37.7135C53.4021 37.0931 72.5371 -15.916 88.0915 9.23596C93.7797 18.4339 88.8931 24.7582 90.2883 33.5573C90.8181 36.898 97.034 33.0261 100.394 32.6337C109.79 31.5364 166.466 20.7134 160.805 43.7169C156.651 60.5994 123.431 55.3648 128.293 38.3292C141.345 -7.39833 269.869 -8.09711 291.076 36.482',
-				translateValues: [220, 415]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['w', 'we ', 'way'],
-		lines: [
-			{
-				path: 'M0.232654 5.81272C-0.890153 4.68991 2.34915 8.20437 3.60108 9.18114C22.2482 23.73 46.0876 41.5951 70.9695 40.7925C91.3189 40.136 110.284 28.9429 127.844 19.8046C135.381 15.8825 142.577 12.2824 148.314 5.94228C149.089 5.08503 154.476 -1.31058 151.682 0.241871',
-				translateValues: [260, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['we have the'],
-		lines: [
-			{
-				path: 'M1.42296 1.21652C-2.92185 0.201698 3.90493 3.72611 5.66717 5.57836C13.0992 13.39 22.0672 20.8371 30.8495 26.991C53.2834 42.7108 81.1296 58.5924 109.792 52.5672C129.927 48.3346 183.997 1.85292 193.827 3.00091C203.767 4.16168 227.799 48.2054 231.176 48.2054C237.106 48.2054 268.06 1.5955 277.014 0.0269309C279.748 -0.452021 277.113 5.58012 277.014 8.35407C276.792 14.5689 276.335 20.7746 276.165 26.991C275.602 47.4961 277.216 67.8229 277.216 88.3403C277.216 131.854 276.713 161.159 279.136 233.3',
-				translateValues: [200, 410]
-			}
-		]
-	},
-	{
-		letterGroupings: ['wt'],
-		specialOutlineMeanings: ['what'],
-		lines: [
-			{
-				path: 'M0 3.05263C7.9168 5.69156 18.7372 23.3454 27.4055 29.1242C59.4052 50.4573 106.219 37.7566 134.602 17.4829C139.445 14.0239 141.485 5.60214 145.758 2.93137C155.026 -2.86116 188.372 1.71874 200.084 1.71874C205.494 1.71874 223.362 -0.914892 216.819 4.26526',
-				translateValues: [240, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['where'],
-		lines: [
-			{
-				path: 'M0.721237 8.32714C-2.50512 8.32714 5.97501 12.0954 8.78524 13.6803C28.2797 24.675 49.7795 32.4782 70.9629 39.4052C122.551 56.2747 168.282 49.1378 219.935 34.2008C244.443 27.1135 266.03 18.1091 286.993 3.42008C288.666 2.24807 294.341 0.766786 289.964 0',
-				translateValues: [220, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['which'],
-		lines: [
-			{
-				path: 'M0 1.03017C27.7457 30.1929 75.0055 41.6807 112.606 28.5071C125.937 23.8366 139.041 17.9859 152.028 12.4912C157.262 10.2769 161.955 7.27726 166.917 4.55663C169.355 3.21945 175.871 -1.67439 174.256 0.589362C170.264 6.18393 145.344 14.4005 151.399 25.7154C155.601 33.5675 167.272 31.5587 174.046 29.9765C175.777 29.5723 186.024 26.2231 181.805 27.3317',
-				translateValues: [230, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['with'],
-		lines: [
-			{
-				path: 'M0.0439908 3.90809C-0.89969 3.90809 13.6273 19.2416 15.9598 21.0085C30.7814 32.2363 47.3436 43.328 66.4659 45.0977C98.7453 48.0851 135.944 41.8712 163.234 23.2389C170.616 18.1991 182.354 2.73335 190.821 0.190613C197.333 -1.76481 183.793 11.8327 180.423 17.7371C177.118 23.5261 170.878 45.7654 165.144 47.7743',
-				translateValues: [260, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['with the'],
-		lines: [
-			{
-				path: 'M0 7.61718C26.2444 19.877 43.9883 39.94 75.7592 29.922C91.0771 25.0919 104.698 18.9881 117.989 9.99636C118.921 9.36594 131.23 -0.651063 132.207 0.033535C133.376 0.852549 117.75 23.369 116.504 25.1637C114.647 27.8378 107.277 31.6709 110.349 32.7473C131.621 40.2 164.363 23.4915 187.594 26.2045C192.747 26.8063 185.614 36.3918 184.835 41.5205C182.224 58.7217 183.336 76.0008 181.44 93.2678C178.348 121.436 195.548 163.214 182.289 187.989',
-				translateValues: [250, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['within'],
-		lines: [
-			{
-				path: 'M0 12.1856C0.188632 11.9874 0.341777 11.7479 0.565895 11.5908C2.45377 10.268 8.91899 15.2129 10.4691 16.1509C34.406 30.6367 61.321 46.2568 89.9773 48.6664C128.183 51.879 158.319 37.1688 187.877 15.5561C192.36 12.2786 200.163 1.48564 205.137 0.0914455C209.123 -1.02595 203.968 8.39301 202.307 12.1856C198.167 21.6418 192.706 32.4086 187.028 40.9341C185.9 42.6286 184.482 46.4855 184.482 46.4855C184.482 46.4855 192.634 36.9781 198.346 35.9775C207.271 34.4141 220.851 36.0093 225.792 45.0977C235.541 63.0301 229.521 94.0993 227.773 113.697C227.451 117.307 227.49 116.502 227.49 115.085',
-				translateValues: [240, 240]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['without'],
-		lines: [
-			{
-				path: 'M0 6.05342C0 13.7078 15.061 20.5272 19.9478 25.2356C48.6242 52.8648 91.237 52.1248 120.748 26.2765C127.506 20.3573 133.437 8.01674 140.059 3.37684C140.313 3.19854 140.483 8.39911 140.483 9.32479C140.483 14.4589 140.663 20.4894 145.152 23.8973C168.261 41.4392 173.541 16.5286 175.498 1.44375C175.935 -1.92483 182.289 1.64404 185.684 1.74115C198.324 2.10265 210.84 2.41067 223.458 2.93074C225.775 3.02626 239.755 2.55954 239.373 3.22814',
-				translateValues: [230, 290]
-			}
-		]
-	},
-	{
-		letterGroupings: ['wts'],
-		specialOutlineMeanings: ['witness'],
-		lines: [
-			{
-				path: 'M0.20484 24.76C-1.51707 24.3578 8.13023 36.6671 9.96652 39.3325C17.5765 50.3782 32.4454 55.6724 45.4057 54.9459C66.1439 53.7833 75.5868 36.7797 91.6676 28.3287C105.227 21.2028 139.291 28.3391 155.543 28.0313C161.006 27.9279 166.353 27.2538 171.671 25.9496C178.766 24.2095 204.913 10.7685 189.284 2.30643C175.609 -5.09832 142.043 6.45831 172.944 20.8938',
-				translateValues: [270, 280]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['witnesses'],
-		lines: [
-			{
-				path: 'M1.80305 21.6936C-6.28115 13.1965 14.9364 30.4556 23.6112 36.5341C41.3228 48.9449 66.3417 47.9866 85.8904 40.6483C96.3399 36.7258 98.6126 27.3257 105.602 22.4283C124.97 8.85629 175.123 33.3772 203.529 11.2612C203.614 11.1948 204.334 10.0172 204.368 9.93876C215.569 -15.8511 156.376 16.6978 189.689 16.6978',
-				translateValues: [210, 270]
-			},
-			{
-				path: 'M0.611141 15.6652C0.611141 26.064 31.2159 21.5128 26.4035 8.0245C18.5055 -14.1125 -10.1633 16.3124 3.75655 18.7508',
-				translateValues: [450, 260]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['worthwhile'],
-		lines: [
-			{
-				path: 'M0.566303 0C-1.93842 0.585032 4.54444 3.30782 6.79115 4.5601C14.8861 9.07214 22.6914 14.2966 30.8417 18.6369C53.2094 30.5487 75.1802 38.1457 100.164 42.2305C131.822 47.4067 168.248 39.0477 193.253 18.2404C197.542 14.6721 204.225 3.29352 209.664 1.38786C212.424 0.421013 212.658 6.43238 214.475 8.72367C233.302 32.4702 269.409 36.3446 292.568 16.0595C294.271 14.5677 303.613 1.78439 304.169 1.78439C305.753 1.78439 304.781 22.4226 304.735 24.1884C303.46 73.323 301.067 109.182 301.067 158.365C301.067 165.608 303.32 194.3 303.32 194.3C303.32 194.3 303.209 188.064 303.603 184.981C305.405 170.879 310.464 155.213 315.77 141.958C330.675 104.72 376.596 31.3667 422.027 39.3252',
-				translateValues: [160, 270]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['x', 'accident'],
-		lines: [
-			{
-				path: 'M0 0C3.98232 1.32744 30.7045 32.5453 30.7045 39.1255',
-				translateValues: [330, 410]
-			},
-			{
-				path: 'M0.0252486 47.7233C-0.535982 48.0974 8.4237 36.6571 10.5192 34.6382C19.2998 26.1788 28.717 18.6191 36.8188 9.50464C37.9764 8.20233 48.0929 -2.55556 47.3127 0.565374',
-				translateValues: [320, 410]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['y', 'your', "you're"],
-		lines: [
-			{
-				path: 'M0 4.79352C5.40523 12.9014 -8.68015 55.1795 13.3441 56.0972C30.5332 56.8134 33.9433 15.6662 33.9433 5.18219C33.9433 4.92405 33.9609 0 33.8138 0C32.7125 0 36.852 35.6286 37.5709 40.4211C38.2344 44.8447 42.4834 57.6201 41.0688 53.3765',
-				translateValues: [300, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['yesterday'],
-		lines: [
-			{
-				path: 'M0 1.78348C11.9178 14.3099 -5.74169 57.528 26.3141 54.7203C42.7267 53.2827 40.7784 22.1491 41.5933 11.3002C41.8668 7.65841 42.6251 -2.20069 40.1078 0.44519C36.7141 4.01218 44.5202 38.5553 47.7474 44.3114C56.3881 59.7233 84.0981 34.9118 60.0556 31.672C54.827 30.9674 49.51 31.2052 44.352 30.4824C38.055 29.5999 56.9756 28.903 63.2387 27.8058C73.477 26.0123 83.6621 24.6647 94.0093 23.9396C132.141 21.2677 169.808 21.5347 207.966 22.75C222.068 23.1992 236.059 23.4935 250.196 23.4935C253.091 23.4935 266.304 23.331 263.778 21.5604',
-				translateValues: [230, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['you'],
-		lines: [
-			{
-				path: 'M3.87206 3.28436C-5.32397 22.6157 1.32953 60.5807 29.5495 55.4777C51.1043 51.5799 62.0743 22.2088 58.6224 1.64867C58.214 -0.783924 58.5425 -0.116558 57.9857 1.05387',
-				translateValues: [300, 400]
-			}
-		]
-	},
-	{
-		specialOutlineMeanings: ['z'],
-		lines: [
-			{
-				path: 'M10.8583 1.55817C6.98705 2.8486 18.7223 6.93598 18.6316 11.0157C17.8684 45.3578 -26.4321 1.95598 24.332 0.00351658C27.3101 -0.111027 29.9287 2.59448 31.328 4.92659C39.8222 19.0837 26.2393 38.6317 17.9838 49.3638C14.0518 54.4755 3.12713 69.8157 4.89881 63.6149',
-				translateValues: [320, 410]
-			}
-		]
-	}
-];
