+++ src/learn-content/connecting-letters.svx
@@ -3,16 +3,17 @@ title: Connecting letters
 description: How and why to join letters together in Teeline shorthand
 ---
 
-<script>
+<script lang="ts">
+    import type { OutlineObject } from '$lib/types.ts'
     import OutlineCardAnimated from '$lib/cards/OutlineCardAnimated.svelte'
     import WordDisemvoweler from '$lib/disemvoweler/wordDisemvoweler.svelte'
     import WordConverter from '$lib/wordConverter.svelte'
     import WordToOutline from '$lib/wordToOutline.svelte'
 
-    import { outlines } from '../data/outlines/outlines.ts'
+    import outlines from '../data/outlines.json'
 	import { findMatchingSpecialOutline } from '../scripts/helpers.ts'
 
-	const conferenceOutline = findMatchingSpecialOutline('conference', outlines);
+	const conferenceOutline: OutlineObject = findMatchingSpecialOutline('conference', outlines);
 </script> 
 
 The full process going on in the shorthand writer's brain is as follows:
