+++ src/learn-content/special-outlines.svx
@@ -5,7 +5,7 @@ descriptions: When Teeline shorthand needs its own shorthand, look no further th
 
 <script>
     import OutlineCardsSideBySide from '$lib/cards/OutlineCardsSideBySide.svelte'
-    import { outlines } from '../data/outlines/outlines.ts'
+    import outlines from '../data/outlines.json'
 	import { findMatchingSpecialOutline } from '../scripts/helpers.ts'
 
 	const dOutline = findMatchingSpecialOutline('ladies and gentlemen', outlines);
